### 2021-02-12 #3035 @whedon
[PRE REVIEW]: Metatheory.jl: Fast and Elegant Algebraic Computation in Julia with Extensible Equality Saturation
**Submitting author:** @0x0f0f0f (<a href="http://orcid.org/0000-0002-8122-9469">Alessandro Cheli</a>)
**Repository:** <a href="https://github.com/0x0f0f0f/Metatheory.jl" target ="_blank">https://github.com/0x0f0f0f/Metatheory.jl</a>
**Version:** v0.1.0
**Reviewers:** @mwillsey, @jpfairbanks, philzook58

Thanks for submitting your paper to JOSS @ 0x0f0f0f. **Currently, there isn't an JOSS editor assigned** to your paper.

@ 0x0f0f0f if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). In addition, [this list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list).
