+++ paper/paper.bib
@@ -16,6 +16,16 @@
   url = {https://github.com/thautwarm/MLStyle.jl}
 }
 
+
+@misc{metatheory,
+  author = {{Cheli}, Alessandro},
+  title = {Metatheory.jl},
+  year = {2021},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  url = {https://github.com/0x0f0f0f/Metatheory.jl}
+}
+
 @misc{matchcore,
   author = {{Zhao}, T and {Carlsson}, K},
   title = {MatchCore.jl: A minimal implementation of optimized pattern matching.},
@@ -88,7 +98,8 @@
 @article{yang2021equality,
   title={Equality Saturation for Tensor Graph Superoptimization},
   author={Yang, Yichen and Phothilimtha, Phitchaya Mangpo and Wang, Yisu Remy and Willsey, Max and Roy, Sudip and Pienaar, Jacques},
-  journal={arXiv preprint arXiv:2101.01332}
+  journal={arXiv preprint arXiv:2101.01332},
+  year={2021}
 }
 
 @article{nelson1980fast,
