### 2021-02-12 c68c7cd
README citation to egg
