+++ paper/paper.bib
@@ -36,14 +36,22 @@
 }
 
 @article{egg,
-  title={egg: fast and extensible equality saturation},
-  author={Willsey, Max and Nandi, Chandrakana and Wang, Yisu Remy and Flatt, Oliver and Tatlock, Zachary and Panchekha, Pavel},
-  journal={Proceedings of the ACM on Programming Languages},
-  volume={5},
-  number={POPL},
-  pages={1--29},
-  year={2021},
-  publisher={ACM New York, NY, USA}
+author = {Willsey, Max and Nandi, Chandrakana and Wang, Yisu Remy and Flatt, Oliver and Tatlock, Zachary and Panchekha, Pavel},
+title = {Egg: Fast and Extensible Equality Saturation},
+year = {2021},
+issue_date = {January 2021},
+publisher = {Association for Computing Machinery},
+address = {New York, NY, USA},
+volume = {5},
+number = {POPL},
+url = {https://doi.org/10.1145/3434304},
+doi = {10.1145/3434304},
+abstract = {An e-graph efficiently represents a congruence relation over many expressions. Although they were originally developed in the late 1970s for use in automated theorem provers, a more recent technique known as equality saturation repurposes e-graphs to implement state-of-the-art, rewrite-driven compiler optimizations and program synthesizers. However, e-graphs remain unspecialized for this newer use case. Equality saturation workloads exhibit distinct characteristics and often require ad-hoc e-graph extensions to incorporate transformations beyond purely syntactic rewrites.  This work contributes two techniques that make e-graphs fast and extensible, specializing them to equality saturation. A new amortized invariant restoration technique called rebuilding takes advantage of equality saturation's distinct workload, providing asymptotic speedups over current techniques in practice. A general mechanism called e-class analyses integrates domain-specific analyses into the e-graph, reducing the need for ad hoc manipulation.  We implemented these techniques in a new open-source library called egg. Our case studies on three previously published applications of equality saturation highlight how egg's performance and flexibility enable state-of-the-art results across diverse domains.},
+journal = {Proc. ACM Program. Lang.},
+month = jan,
+articleno = {23},
+numpages = {29},
+keywords = {e-graphs, equality saturation}
 }
 
 @article{bezanson2017julia,
@@ -54,8 +62,12 @@
   number={1},
   pages={65--98},
   year={2017},
+  doi = {10.1137/141000671},
   publisher={SIAM}
 }
+
+
+
 
 @article{belyakova2020world,
   title={World Age in Julia},
@@ -75,15 +87,44 @@
 }
 
 @article{panchekha2015automatically,
-  title={Automatically improving accuracy for floating point expressions},
-  author={Panchekha, Pavel and Sanchez-Stern, Alex and Wilcox, James R and Tatlock, Zachary},
-  journal={ACM SIGPLAN Notices},
-  volume={50},
-  number={6},
-  pages={1--11},
-  year={2015},
-  publisher={ACM New York, NY, USA}
+author = {Panchekha, Pavel and Sanchez-Stern, Alex and Wilcox, James R. and Tatlock, Zachary},
+title = {Automatically Improving Accuracy for Floating Point Expressions},
+year = {2015},
+issue_date = {June 2015},
+publisher = {Association for Computing Machinery},
+address = {New York, NY, USA},
+volume = {50},
+number = {6},
+issn = {0362-1340},
+url = {https://doi.org/10.1145/2813885.2737959},
+doi = {10.1145/2813885.2737959},
[…]
