### 2021-02-14 438afda
fixed minor bugs, right hand sydes in dynamic rules are now based on RuntimeGeneratedFunctions and are much faster.
