+++ src/EGraphs/extraction.jl
@@ -35,7 +35,6 @@ modify!(analysis::ExtractionAnalysis, G::EGraph, id::Int64) = nothing
 
 function rec_extract(G::EGraph, data, id::Int64)
     # println(data[id])
-    # TODO in Julia 1.5 this is sometimes nothing. WTF.
     (cn, ck) = data[id]
     !(cn isa Expr) && return cn
 
