+++ TODO.md
@@ -1,13 +1,13 @@
 TODO:
 * Extend regular binary operators from a theory `src/operators.jl`
 * `@judgment` for natural deduction syntax style like racket's redex
-* Abstract interface to allow interchangeable solvers????
 * FIXME allow ∀ and ∃ as symbols: julia parser issue
-* E-Graph extractor from cost function
 * Check if matching in ematch.jl can be made concurrent
 * Try getting MatchCore to work with egraphs
 
 DONE:
+* E-Graph extractor from cost function
+* Abstract interface to allow interchangeable solvers????
 * Equational theories with e-graphs
 * Deal WITH N-ARITY OF * AND + - `binarize` hack in `src/reduce.jl`
 * FIXED runtime rewrite
