### 2021-02-14 a273d9e
adapted theory definition with algebraic structures convenience functions for the e-graph backend, adapted and simplified tests, identified a RGF bug
