+++ src/Library/Library.jl
@@ -1,71 +1,80 @@
 # constructs a semantic theory about a commutative monoid
 # A monoid whose operation is commutative is called a
 # commutative monoid (or, less commonly, an abelian monoid).
-# TODO test
 
-# TODO split everything up. Use auto ACD compiler for matchcore backend
-macro commutative_monoid(t, op, id, func)
-	zop = Meta.quot(op)
-   quote
-	   @theory begin
-			($op)($id, a) => a
-			($op)(a, $id) => a
-			# closure
-			($op)(a::$t, b::$t) |> ($func)(a,b)
-			# associativity reductions implying commutativity
-			($op)(a::$t, ($op)(b::$t, c)) |>
-				(z = ($func)(a, b); Expr(:call, ($zop), z, c))
+# TODO THIS WORKS CORRECTLY ONLY IN E-GRAPHS
 
-			($op)(($op)(a::$t, c), b::$t) |>
-				(z = ($func)(a, b); Expr(:call, ($zop), z, c))
+commutativity(op) = :( ($op)(a, b) => ($op)(b, a) ) |> Rule
+associativity(op) = :( ($op)(a, $(op)(b,c)) => ($op)($(op)(a,b), c) ) |> Rule
 
-			($op)(a::$t, ($op)(c, b::$t)) |>
-				(z = ($func)(a, b); Expr(:call, ($zop), z, c))
+identity_left(op, id) = :( ($op)($id, a) => a ) |> Rule
+identity_right(op, id) = :( ($op)(a, $id) => a ) |> Rule
 
-			($op)(($op)(c, a::$t), b::$t) |>
-				(z = ($func)(a, b); Expr(:call, ($zop), z, c))
-	   end
-   end
+inverse_left(op, id, invop) = :( ($op)(($invop)(a), a) => $id ) |> Rule
+inverse_right(op, id, invop) = :( ($op)(a, ($invop)(a)) => $id ) |> Rule
+
+function commutative_monoid(op, id)
+	@assert Base.isbinaryoperator(op)
+	[commutativity(op), associativity(op), identity_left(op, id)]
 end
-
+macro commutative_monoid(op, id) commutative_monoid(op, id) end
 
 # constructs a semantic theory about a an abelian group
 # The definition of a group does not require that a ⋅ b = b ⋅ a
 # for all elements a and b in G. If this additional condition holds,
 # then the operation is said to be commutative, and the group is called an abelian group.
-# TODO test
-macro abelian_group(t, op, id, invop, func)
-	zop = Meta.quot(op)
-   quote
-	   @theory begin
-			# identity element
-			($op)($id, a) => a
-			($op)(a, $id) => a
-			# inversibility
-			($op)(a, ($invop)(a)) => $id
-			($op)(($invop)(a), a) => $id
-			# closure
-			($op)(a::$t, b::$t) |> ($func)(a,b)
-			# inverse
-			# associativity reductions
-			($op)(a::$t, ($op)(b::$t, c)) |> (z = ($func)(a, b); Expr(:call, ($zop), z, c))
-			($op)(($op)(a::$t, c), b::$t) |> (z = ($func)(a, b); Expr(:call, ($zop), z, c))
-			($op)(a::$t, ($op)(c, b::$t)) |> (z = ($func)(a, b); Expr(:call, ($zop), z, c))
-			($op)(($op)(c, a::$t), b::$t) |> (z = ($func)(a, b); Expr(:call, ($zop), z, c))
-	   end
-   end
+function commutative_group(op, id, invop)
+	@assert Base.isbinaryoperator(op)
+	# @assert Base.isunaryoperator(invop)
+	commutative_monoid(op, id) ∪ [inverse_right(op, id, invop)]
 end
-
[…]
