### 2021-02-14 b6110c3
tidying up the library
