+++ docs/make.jl
@@ -1,0 +1,5 @@
+using Documenter, Metatheory
+
+makedocs(modules = [Metatheory], sitename = "Metatheory.jl")
+
+deploydocs(repo = "github.com/0x0f0f0f/Metatheory.jl.git")
