+++ docs/src/index.md
@@ -1,0 +1,7 @@
+# Example
+
+Example Julia package repo.
+
+```@autodocs
+Modules = [Metatheory]
+```
