+++ docs/src/api.md
@@ -1,0 +1,5 @@
+# API Documentation
+
+```@autodocs
+Modules = [Metatheory, Metatheory.Library]
+```
