+++ test/runtests.jl
@@ -6,6 +6,8 @@ using Test
 
 include("../src/util.jl")
 
+Metatheory.init(@__MODULE__)
+
 include("test_theories.jl")
 include("test_reductions.jl")
 include("test_egraphs.jl")
@@ -14,6 +16,7 @@ include("test_ematch_assertions.jl")
 include("test_analysis.jl")
 include("test_dynamic_ematch.jl")
 include("test_extraction.jl")
+include("test_mu.jl")
 include("test_while_interpreter.jl")
 
 # exported consistency test

[…]
