+++ test/test_while_superinterpreter.jl
@@ -1,0 +1,113 @@
+
+## Turing Complete Interpreter
+### A Very Tiny Turing Complete Programming Language defined with denotational semantics
+
+# semantica dalle dispense degano
+
+const Mem = Base.ImmutableDict{Symbol, Union{Bool, Int}}
+
+read_mem = @theory begin
+	(v::Symbol, σ::Mem) |> σ[v]
+end
+
+@testset "Reading Memory" begin
+	@test true == areequal(read_mem, :((x), $(Mem(:x, 2))), 2, mod=@__MODULE__)
+end
+
+arithm_rules = @theory begin
+	(a + b, σ) 		  	=> (a, σ) + (b, σ)
+	(a * b, σ) 		  	=> (a, σ) * (b, σ)
+	(a - b, σ) 		  	=> (a, σ) - (b, σ)
+	(n::Int, σ) 		=> n
+	(a::Int + b::Int) 	|> a + b
+	(a::Int * b::Int) 	|> a * b
+	(a::Int - b::Int) 	|> a - b
+end
+
+
+@testset "Arithmetic" begin
+	@test areequal(read_mem ∪ arithm_rules,
+		:((2 + 3), $(Mem())), 5, mod=@__MODULE__)
+	@test areequal(read_mem ∪ arithm_rules,
+		:((2 + x), 4; mod=@__MODULE__))
+end
+
+# don't need to access memory
+bool_rules = @theory begin
+	(a::Bool ∨ b::Bool, σ) 	|> (a || b)
+	(a::Bool ∧ b::Bool, σ) 	|> (a && b)
+	(a::Bool ∨ b::Bool) 	|> (a || b)
+	(a::Bool ∧ b::Bool) 	|> (a && b)
+	(a::Bool, σ) 			|> a
+	(¬a::Bool, σ) 			|> !a
+	¬a::Bool 				|> !a
+	(a::Int < b::Int, σ) |> (a < b)
+	(a::Int < b::Int) 	 |> (a < b)
+	¬(a, σ) => (¬a, σ)
+
+	# (¬b, σ) |> !eval_bool(b, σ)
+	(a < b, σ) => (a, σ) < (b, σ)
+	(a ∨ b, σ) => (a, σ) ∨ (b, σ)
+	(a ∧ b, σ) => (a, σ) ∧ (b, σ)
+end
+
+t = read_mem ∪ arithm_rules ∪ bool_rules
+
+@testset "Booleans" begin
+	@test areequal(t, :((false ∨ false), $(Mem())),
+		false; mod=@__MODULE__)
+	@test areequal(t, :((false ∨ false) ∨ ¬(false ∨ false), $(Mem(:x, 2))),
+		true; mod=@__MODULE__)
+	@test areequal(t, :((2 < 3) ∧ (3 < 4), $(Mem(:x, 2))),
+		true; mod=@__MODULE__)
+	@test areequal(t, :((2 < x) ∨ ¬(3 < 4), $(Mem(:x, 2))),
+		false; mod=@__MODULE__)
+	@test  areequal(t, :((2 < x) ∨ ¬(3 < 4), $(Mem(:x, 4))),
+		true; mod=@__MODULE__)
+end
+
+exit(0)
+
+if_rules = @theory begin
+	(if guard; t end, σ) => (if guard; t else skip end, σ)
+	(if guard; t else f end, σ) |>
+		(eval_bool(guard, σ) ? :($t, $σ) : :($f, $σ))
+end
+
+eval_if(ex::Expr, mem::Mem) = (@rewriter(read_mem ∪ arithm_rules ∪ if_rules, :inner))(:($ex, $mem))
+
+@testset "If Semantics" begin
+	@test 2 == eval_if(:(if true x else 0 end), Mem(:x => 2))
[…]
