+++ src/Metatheory.jl
@@ -1,10 +1,16 @@
 module Metatheory
 
 using RuntimeGeneratedFunctions
+using Base.Meta
+
 RuntimeGeneratedFunctions.init(@__MODULE__)
 
 
-include("util.jl")
+include("Util/Util.jl")
+using .Util
+
+
+
 include("rule.jl")
 include("theory.jl")
 include("matchcore_compiler.jl")
@@ -12,6 +18,7 @@ include("rewrite.jl")
 include("match.jl")
 include("EGraphs/EGraphs.jl")
 
+export @metatheory_init
 
 include("Library/Library.jl")
 export Library
