+++ src/EGraphs/schedulers/abstractscheduler.jl
@@ -1,1 +1,0 @@
-abstract type AbstractScheduler end
