+++ src/Metatheory.jl
@@ -5,10 +5,9 @@ using Base.Meta
 
 RuntimeGeneratedFunctions.init(@__MODULE__)
 
-
 include("Util/Util.jl")
 using .Util
-
+export Util
 
 
 include("rule.jl")
