+++ docs/src/api.md
@@ -1,5 +1,0 @@
-# API Documentation
-
-```@autodocs
-Modules = [Metatheory, Metatheory.Library]
-```
