+++ docs/src/modules/metatheory.md
@@ -1,0 +1,6 @@
+# Metatheory
+
+
+```@autodocs
+Modules = [Metatheory]
+```
