+++ docs/src/modules/util.md
@@ -1,0 +1,15 @@
+# Metatheory.Util
+
+```@meta
+CurrentModule = Metatheory
+DocTestSetup  = quote
+    using MyPackage
+end
+DocTestFilters = [r"Stacktrace:[\s\S]+"]
+EditURL = "link/to/source/file"
+```
+
+
+```@autodocs
+Modules = [Util]
+```
