+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"api/#API-Documentation","page":"API Documentation","title":"API Documentation","text":"","category":"section"},{"location":"api/","page":"API Documentation","title":"API Documentation","text":"Modules = [Metatheory, Metatheory.Library]","category":"page"},{"location":"api/#Metatheory.RhsFunCache","page":"API Documentation","title":"Metatheory.RhsFunCache","text":"Type representing a cache of RuntimeGeneratedFunctions corresponding to right hand sides of dynamic rules\n\n\n\n\n\n","category":"type"},{"location":"api/#Metatheory.eval_types_in_assertions-Tuple{Any,Module}","page":"API Documentation","title":"Metatheory.eval_types_in_assertions","text":"When creating a theory, type assertions in the left hand contain symbols. We want to replace the type symbols with the real type values, to fully support the subtyping mechanism during pattern matching.\n\n\n\n\n\n","category":"method"},{"location":"api/#Metatheory.genrhsfun-Tuple{Rule,Module}","page":"API Documentation","title":"Metatheory.genrhsfun","text":"Generates a tuple containing the list of formal parameters (Symbols) and the RuntimeGeneratedFunction corresponding to the right hand side of a :dynamic Rule.\n\n\n\n\n\n","category":"method"},{"location":"#Metatheory.jl","page":"Metatheory.jl","title":"Metatheory.jl","text":"","category":"section"},{"location":"","page":"Metatheory.jl","title":"Metatheory.jl","text":"Docs will appear here some time in the near future.","category":"page"}]
+[{"location":"modules/metatheory/#Metatheory","page":"Metatheory","title":"Metatheory","text":"","category":"section"},{"location":"modules/metatheory/","page":"Metatheory","title":"Metatheory","text":"Modules = [Metatheory]","category":"page"},{"location":"modules/metatheory/#Metatheory.RhsFunCache","page":"Metatheory","title":"Metatheory.RhsFunCache","text":"Type representing a cache of RuntimeGeneratedFunctions corresponding to right hand sides of dynamic rules\n\n\n\n\n\n","category":"type"},{"location":"modules/metatheory/#Metatheory.eval_types_in_assertions-Tuple{Any,Module}","page":"Metatheory","title":"Metatheory.eval_types_in_assertions","text":"When creating a theory, type assertions in the left hand contain symbols. We want to replace the type symbols with the real type values, to fully support the subtyping mechanism during pattern matching.\n\n\n\n\n\n","category":"method"},{"location":"modules/metatheory/#Metatheory.genrhsfun-Tuple{Rule,Module}","page":"Metatheory","title":"Metatheory.genrhsfun","text":"Generates a tuple containing the list of formal parameters (Symbols) and the RuntimeGeneratedFunction corresponding to the right hand side of a :dynamic Rule.\n\n\n\n\n\n","category":"method"},{"location":"modules/util/#Metatheory.Util","page":"Metatheory.Util","title":"Metatheory.Util","text":"","category":"section"},{"location":"modules/util/","page":"Metatheory.Util","title":"Metatheory.Util","text":"CurrentModule = Metatheory\nDocTestSetup  = quote\n    using MyPackage\nend\nDocTestFilters = [r\"Stacktrace:[\\s\\S]+\"]\nEditURL = \"link/to/source/file\"","category":"page"},{"location":"modules/util/","page":"Metatheory.Util","title":"Metatheory.Util",
[…]
