### 2021-02-17 a8752dd
module refactoring. preparing for first release. docs
