+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"modules/metatheory/#Metatheory","page":"Metatheory","title":"Metatheory","text":"","category":"section"},{"location":"modules/metatheory/","page":"Metatheory","title":"Metatheory","text":"Modules = [Metatheory]","category":"page"},{"location":"modules/metatheory/#Metatheory.RhsFunCache","page":"Metatheory","title":"Metatheory.RhsFunCache","text":"Type representing a cache of RuntimeGeneratedFunctions corresponding to right hand sides of dynamic rules\n\n\n\n\n\n","category":"type"},{"location":"modules/metatheory/#Metatheory.eval_types_in_assertions-Tuple{Any,Module}","page":"Metatheory","title":"Metatheory.eval_types_in_assertions","text":"When creating a theory, type assertions in the left hand contain symbols. We want to replace the type symbols with the real type values, to fully support the subtyping mechanism during pattern matching.\n\n\n\n\n\n","category":"method"},{"location":"modules/metatheory/#Metatheory.genrhsfun-Tuple{Rule,Module}","page":"Metatheory","title":"Metatheory.genrhsfun","text":"Generates a tuple containing the list of formal parameters (Symbols) and the RuntimeGeneratedFunction corresponding to the right hand side of a :dynamic Rule.\n\n\n\n\n\n","category":"method"},{"location":"modules/util/#Metatheory.Util","page":"Metatheory.Util","title":"Metatheory.Util","text":"","category":"section"},{"location":"modules/util/","page":"Metatheory.Util","title":"Metatheory.Util","text":"CurrentModule = Metatheory\nDocTestSetup  = quote\n    using MyPackage\nend\nDocTestFilters = [r\"Stacktrace:[\\s\\S]+\"]\nEditURL = \"link/to/source/file\"","category":"page"},{"location":"modules/util/","page":"Metatheory.Util","title":"Metatheory.Util","text":"Modules = [Util]","category":"page"},{"location":"modules/util/#Metatheory.Util","page":"Metatheory.Util","title":"Metatheory.Util","text":"Definitions of various utility functions for metaprogramming\n\n\n\n\n\n","category":"module"},{"location":"modules/util/#Metatheory.Util.amp-Tuple{Any}","page":"Metatheory.Util","title":"Metatheory.Util.amp","text":"Add a & expression\n\n\n\n\n\n","category":"method"},{"location":"modules/util/#Metatheory.Util.bf_walk!-Tuple{Any,Any,Vararg{Any,N} where N}","page":"Metatheory.Util","title":"Metatheory.Util.bf_walk!","text":"Breadth First Walk (Tree Prewalk) on expressions mutates expression in-place.\n\n\n\n\n\n","category":"method"},{"location":"modules/util/#Metatheory.Util.bf_walk-Tuple{Any,Any,Vararg{Any,N} where N}","page":"Metatheory.Util","title":"Metatheory.Util.bf_walk","text":"Breadth First Walk (Tree Prewalk) on expressions. Does not mutate expressions.\n\n\n\n\n\n","category":"method"},{"location":"modules/util/#Metatheory.Util.binarize!-Tuple{Any,Symbol}","page":"Metatheory.Util","title":"Metatheory.Util.binarize!","text":"HARD FIX of n-arity of operators in Expr trees\n\n\n\n\n\n","category":"method"},{"location":"modules/util/#Metatheory.Util.block-Tuple","page":"Metatheory.Util","title":"Metatheory.Util.block","text":"Make a block expression from an array of exprs\n\n\n\n\n\n","category":"method"},{"location":"module
[…]
