### 2021-02-17 #3035 @Kevin-Mattheus-Moerman
> I don't understand why whedon does not list any commit. I have pushed 90 so far.

@0x0f0f0f I believe this feature (for whedon to scan the repository) is relatively new and might not yet be 100% accurate. I'll ping @openjournals/dev to see if they can check this out.
