### 2021-02-17 #3035 @arfon
> @0x0f0f0f I believe this feature (for whedon to scan the repository) is relatively new and might not yet be 100% accurate. I'll ping @openjournals/dev to see if they can check this out.

Not sure what is going on here sorry. We simply run `gitinspector` (https://github.com/openjournals/whedon-api/blob/master/workers.rb#L379) so perhaps it is struggling with this repository for some reason.
