+++ .github/workflows/docs.yml
@@ -1,0 +1,24 @@
+name: Docs
+on:
+  pull_request:
+    paths:
+      - "docs/"
+    branches:
+      - master
+  push:
+    branches:
+      - master
+    tags: '*'
+    paths:
+      - "docs/"
+jobs:
+  docs:
+    name: Documentation
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+      - uses: julia-actions/julia-buildpkg@latest
+      - uses: julia-actions/julia-docdeploy@latest
+        env:
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+          DOCUMENTER_KEY: ${{ secrets.DOCUMENTER_KEY }}
