### 2021-02-17 f2b69af
docs and citation
