+++ docs/src/modules/metatheory.md
@@ -1,6 +1,0 @@
-# Metatheory
-
-
-```@autodocs
-Modules = [Metatheory]
-```
