+++ docs/src/modules/egraphs.md
@@ -1,13 +1,0 @@
-# Metatheory.EGraphs
-
-```@meta
-CurrentModule = Metatheory
-DocTestSetup  = quote
-    using Metatheory
-end
-```
-
-
-```@autodocs
-Modules = [EGraphs]
-```
