+++ CITATION.bib
@@ -1,1 +1,8 @@
-TODO :)
+@misc{cheli2021metatheoryjl,
+      title={Metatheory.jl: Fast and Elegant Algebraic Computation in Julia with Extensible Equality Saturation},
+      author={Alessandro Cheli},
+      year={2021},
+      eprint={2102.07888},
+      archivePrefix={arXiv},
+      primaryClass={cs.PL}
+}
