+++ docs/src/modules/util.md
@@ -1,13 +1,0 @@
-# Metatheory.Util
-
-```@meta
-CurrentModule = Metatheory
-DocTestSetup  = quote
-    using Metatheory
-end
-```
-
-
-```@autodocs
-Modules = [Util]
-```
