+++ docs/src/api.md
@@ -1,0 +1,49 @@
+# API Documentation
+
+## Metatheory
+
+```@autodocs
+Modules = [Metatheory]
+```
+
+## Metatheory.EGraphs
+
+```@meta
+CurrentModule = Metatheory
+DocTestSetup  = quote
+    using Metatheory
+end
+```
+
+
+```@autodocs
+Modules = [Metatheory.EGraphs]
+```
+
+## Metatheory.EGraphs.Schedulers
+
+```@meta
+CurrentModule = Metatheory
+DocTestSetup  = quote
+    using Metatheory
+end
+```
+
+
+```@autodocs
+Modules = [Metatheory.EGraphs.Schedulers]
+```
+
+## Metatheory.Util
+
+```@meta
+CurrentModule = Metatheory
+DocTestSetup  = quote
+    using Metatheory
+end
+```
+
+
+```@autodocs
+Modules = [Util]
+```
