+++ docs/src/index.md
@@ -1,3 +1,4 @@
 # Metatheory.jl
 
 Docs will appear here some time in the near future.
+![dragon](./assets/dragon.jpg)
