+++ docs/src/index.md
@@ -1,4 +1,9 @@
 # Metatheory.jl
 
-Docs will appear here some time in the near future.
-![dragon](./assets/dragon.jpg)
+<p align="center">
+<img width="400px" src="https://raw.githubusercontent.com/0x0f0f0f/Metatheory.jl/master/docs/src/assets/dragon.jpg"/>
+</p>
+
+
+**Metatheory.jl** is a general purpose metaprogramming and algebraic computation library for the Julia programming language, designed to take advantage of the powerful reflection capabilities to bridge the gap between symbolic mathematics, abstract interpretation, equational reasoning, optimization, composable compiler transforms, and advanced
+homoiconic pattern matching features.
