+++ docs/Project.toml
@@ -1,4 +1,6 @@
 [deps]
+Calculus = "49dc2e85-a5d0-5ad3-a950-438e2897f1b9"
+DataStructures = "864edb3b-99cc-5e75-8d2d-829cb0a9cfe8"
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
 DocumenterTools = "35a29f4d-8980-5a13-9543-d66fff28ecb8"
 MatchCore = "5dd3f0b1-72a9-48ad-ae6e-79f673da005f"
