+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"api/#API-Documentation","page":"API Documentation","title":"API Documentation","text":"","category":"section"},{"location":"api/#Metatheory","page":"API Documentation","title":"Metatheory","text":"","category":"section"},{"location":"api/","page":"API Documentation","title":"API Documentation","text":"Modules = [Metatheory]","category":"page"},{"location":"api/#Main.Metatheory.Rule","page":"API Documentation","title":"Main.Metatheory.Rule","text":"Structure representing rewrite rules.\n\n\n\n\n\n","category":"type"},{"location":"api/#Main.Metatheory.compile_theory-Tuple{Array{Main.Metatheory.Rule,1},Module}","page":"API Documentation","title":"Main.Metatheory.compile_theory","text":"Compile a theory to a closure that does the pattern matching job Returns a RuntimeGeneratedFunction, which does not use eval and is as fast as a regular Julia anonymous function 🔥\n\n\n\n\n\n","category":"method"},{"location":"api/#Main.Metatheory.@compile_theory-Tuple{Any}","page":"API Documentation","title":"Main.Metatheory.@compile_theory","text":"Compile a theory at runtime to a closure that does the pattern matching job\n\n\n\n\n\n","category":"macro"},{"location":"api/#Metatheory.EGraphs","page":"API Documentation","title":"Metatheory.EGraphs","text":"","category":"section"},{"location":"api/","page":"API Documentation","title":"API Documentation","text":"CurrentModule = Metatheory\nDocTestSetup  = quote\n    using Metatheory\nend","category":"page"},{"location":"api/","page":"API Documentation","title":"API Documentation","text":"Modules = [Metatheory.EGraphs]","category":"page"},{"location":"api/#Main.Metatheory.EGraphs.RhsFunCache","page":"API Documentation","title":"Main.Metatheory.EGraphs.RhsFunCache","text":"Type representing a cache of RuntimeGeneratedFunctions corresponding to right hand sides of dynamic rules\n\n\n\n\n\n","category":"type"},{"location":"api/#Main.Metatheory.EGraphs.eval_types_in_assertions-Tuple{Any,Module}","page":"API Documentation","title":"Main.Metatheory.EGraphs.eval_types_in_assertions","text":"When creating a theory, type assertions in the left hand contain symbols. We want to replace the type symbols with the real type values, to fully support the subtyping mechanism during pattern matching.\n\n\n\n\n\n","category":"method"},{"location":"api/#Main.Metatheory.EGraphs.genrhsfun-Tuple{Main.Metatheory.Rule,Module}","page":"API Documentation","title":"Main.Metatheory.EGraphs.genrhsfun","text":"Generates a tuple containing the list of formal parameters (Symbols) and the RuntimeGeneratedFunction corresponding to the right hand side of a :dynamic Rule.\n\n\n\n\n\n","category":"method"},{"location":"api/#Metatheory.EGraphs.Schedulers","page":"API Documentation","title":"Metatheory.EGraphs.Schedulers","text":"","category":"section"},{"location":"api/","page":"API Documentation","title":"API Documentation","text":"CurrentModule = Metatheory\nDocTestSetup  = quote\n    using Metatheory\nend","category":"page"},{"location":"api/","page":"API Documentation","title":"API Documentation","text":"Modules = [Metatheory.EGraphs.Schedulers]","category":"page"},{"location":"api/#Main.Metatheory.EGraphs.Schedulers.AbstractScheduler","page":"API Documentation","t
[…]
