### 2021-02-18 1a3251b
Redesigned the AbstractAnalysis interface and the analysis mechanism.
