### 2021-02-19 5761ac0
added astsize_inv cost function
