### 2021-02-19 ec71dab
removed superfluous prints
