+++ src/EGraphs/saturation.jl
@@ -3,16 +3,17 @@ const MatchesBuf = Vector{Tuple{Rule, Sub, Int64}}
 using .Schedulers
 
 
-inst(var, G::EGraph, sub::Sub) = haskey(sub, var) ? first(sub[var]) : add!(G, var)
-inst(p::Expr, G::EGraph, sub::Sub) = add!(G, p)
+inst(var, egraph::EGraph, sub::Sub) =
+    haskey(sub, var) ? first(sub[var]) : add!(egraph, var)
+inst(p::Expr, egraph::EGraph, sub::Sub) = add!(egraph, p)
 
-function instantiate(G::EGraph, p, sub::Sub; skip_assert=false)
+function instantiate(egraph::EGraph, p, sub::Sub; skip_assert=false)
     # remove type assertions
     if skip_assert
         p = df_walk( x -> (isexpr(x, :ematch_tassert) ? x.args[1] : x), p; skip_call=true )
     end
 
-    df_walk(inst, p, G, sub; skip_call=true)
+    df_walk(inst, p, egraph, sub; skip_call=true)
 end
 
 function get_actual_param(var, sub::Sub)
@@ -25,7 +26,7 @@ function get_actual_param(var, sub::Sub)
 end
 
 
-function eqsat_step!(G::EGraph, theory::Vector{Rule};
+function eqsat_step!(egraph::EGraph, theory::Vector{Rule};
         scheduler=SimpleScheduler(), rhs_funs=RhsFunCache())
     matches=MatchesBuf()
     EMPTY_DICT = Sub()
@@ -37,13 +38,13 @@ function eqsat_step!(G::EGraph, theory::Vector{Rule};
         shouldskip(scheduler, rule) && continue
 
         if rule.mode == :rewrite || rule.mode == :dynamic
-            for id ∈ keys(G.M)
+            for id ∈ keys(egraph.M)
                 # println(rule.right)
-                for sub in ematch(G, rule.left, id, EMPTY_DICT)
+                for sub in ematch(egraph, rule.left, id, EMPTY_DICT)
                     # display(sub); println()
                     !isempty(sub) && push!(matches, (rule, sub, id))
                 end
-                # for sub in ematch(G, rule.right, id, EMPTY_DICT)
+                # for sub in ematch(egraph, rule.right, id, EMPTY_DICT)
                 #     # display(sub); println()
                 #     !isempty(sub) && push!(matches, (rule, sub, id))
                 # end
@@ -58,30 +59,30 @@ function eqsat_step!(G::EGraph, theory::Vector{Rule};
         writestep!(scheduler, rule)
 
         if rule.mode == :rewrite # symbolic replacement
-            l = instantiate(G,rule.left,sub; skip_assert=true)
-            r = instantiate(G,rule.right,sub)
-            merge!(G,l.id,r.id)
+            l = instantiate(egraph,rule.left,sub; skip_assert=true)
+            r = instantiate(egraph,rule.right,sub)
+            merge!(egraph,l.id,r.id)
         elseif rule.mode == :dynamic # execute the right hand!
-            l = instantiate(G,rule.left,sub; skip_assert=true)
+            l = instantiate(egraph,rule.left,sub; skip_assert=true)
 
             (params, f) = rhs_funs[rule]
             actual_params = params .|> x -> get_actual_param(x, sub)
-            r = addexpr!(G, f(actual_params...))
-            merge!(G,l.id,r.id)
+            r = addexpr!(egraph, f(egraph, actual_params...))
+            merge!(egraph,l.id,r.id)
         else
             error("unsupported rule mode")
         end
     end
 
-    # display(G.parents); println()
-    # display(G.M); println()
[…]
