### 2021-02-21 b3d3b53
Fixed symbol in right hand bug
