### 2021-02-22 68cc2a1
working on lazy analyses algorithm
