+++ .github/workflows/docs.yml
@@ -3,6 +3,7 @@ on:
   pull_request:
     paths:
       - "docs/**"
+      - "src/**"
     branches:
       - master
   push:
