+++ src/rewrite.jl
@@ -1,5 +1,30 @@
 const TIMEOUT = 1000
 
+"""
+This function executes a classical rewriting
+algorithm on a Julia expression `ex`. Classical rewriting
+applies rule in order with a fixed point iteration:
+
+This algorithm heavily relies on
+[RuntimeGeneratedFunctions.jl](https://github.com/SciML/RuntimeGeneratedFunctions.jl)
+and the [MatchCore](https://github.com/SciML/RuntimeGeneratedFunctions.jl)
+pattern matcher.
+**NOTE**: this does not involve the use of [`EGraphs.EGraph`](@ref) or
+equality saturation ([`EGraphs.saturate!`](@ref)).
+When using `rewrite`, be aware of infinite loops:
+Since rules are matched in order in every iteration, it
+is possible that commonly used symbolic rules such as commutativity
+or associativity of operators may cause this algorithm to
+have a cycling computation instantly. This algorithm
+detects cycling computation by keeping an history of hashes,
+and instantly returns when a cycle is detected.
+
+This algorithm is suitable for simple, deterministic symbolic rewrites.
+For more advanced use cases, where it is needed to apply multiple
+rewrites at the same time, or it is known that rules are causing loops,
+please use [`EGraphs.EGraph`](@ref) and
+equality saturation ([`EGraphs.saturate!`](@ref)).
+"""
 function rewrite(ex, theory::Theory;
         __source__=LineNumberNode(0),
         order=:inner,                   # evaluation order
