+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"api/#API-Documentation","page":"API Documentation","title":"API Documentation","text":"","category":"section"},{"location":"api/#Metatheory","page":"API Documentation","title":"Metatheory","text":"","category":"section"},{"location":"api/","page":"API Documentation","title":"API Documentation","text":"Modules = [Metatheory]","category":"page"},{"location":"api/#Main.Metatheory.Rule","page":"API Documentation","title":"Main.Metatheory.Rule","text":"mutable struct Rule\n\nStructure representing rewrite rules.\n\n\n\nFields\n\nleft::Any\nright::Any\nexpr::Expr\nmode::Symbol\n\n\n\n\n\n","category":"type"},{"location":"api/#Main.Metatheory.compile_theory-Tuple{Array{Main.Metatheory.Rule,1},Module}","page":"API Documentation","title":"Main.Metatheory.compile_theory","text":"Compile a theory to a closure that does the pattern matching job Returns a RuntimeGeneratedFunction, which does not use eval and is as fast as a regular Julia anonymous function 🔥\n\n\n\nSignatures\n\ncompile_theory(theory::Array{Main.Metatheory.Rule,1}, mod::Module; __source__) -> RuntimeGeneratedFunctions.RuntimeGeneratedFunction{_A,_B,_C,_D} where _D where _C where _B where _A\n\n\n\n\nMethods\n\ncompile_theory(theory, mod; __source__)\n\ndefined at /home/runner/work/Metatheory.jl/Metatheory.jl/src/matchcore_compiler.jl:61.\n\n\n\n\n\n","category":"method"},{"location":"api/#Main.Metatheory.@compile_theory-Tuple{Any}","page":"API Documentation","title":"Main.Metatheory.@compile_theory","text":"Compile a theory at runtime to a closure that does the pattern matching job\n\n\n\nSignatures\n\n\n\nMethods\n\n\n\n\n\n","category":"macro"},{"location":"api/#Metatheory.EGraphs","page":"API Documentation","title":"Metatheory.EGraphs","text":"","category":"section"},{"location":"api/","page":"API Documentation","title":"API Documentation","text":"CurrentModule = Metatheory\nDocTestSetup  = quote\n    using Metatheory\nend","category":"page"},{"location":"api/","page":"API Documentation","title":"API Documentation","text":"Modules = [Metatheory.EGraphs]","category":"page"},{"location":"api/#Main.Metatheory.EGraphs.AbstractAnalysis","page":"API Documentation","title":"Main.Metatheory.EGraphs.AbstractAnalysis","text":"abstract type AbstractAnalysis\n\n\n\nFields\n\n\n\n\n\n","category":"type"},{"location":"api/#Main.Metatheory.EGraphs.EGraph","page":"API Documentation","title":"Main.Metatheory.EGraphs.EGraph","text":"mutable struct EGraph\n\n\n\nFields\n\nU::DataStructures.IntDisjointSets\nequality relation over e-class ids\nM::Dict{Int64,Array{Any,1}}\nH::Dict{Any,Int64}\nparents::Dict{Int64,Array{Tuple{Any,Int64},1}}\ndirty::Array{Int64,1}\nroot::Int64\nanalyses::Array{Main.Metatheory.EGraphs.AbstractAnalysis,1}\n\n\n\n\n\n","category":"type"},{"location":"api/#Main.Metatheory.EGraphs.Extracti
[…]
