### 2021-02-23 5ecbb06
working lazy analyses! fixes for macro expansion in theories. added tests and few util functions
