+++ src/Metatheory.jl
@@ -19,6 +19,13 @@ include("theory.jl")
 include("matchcore_compiler.jl")
 include("rewrite.jl")
 include("match.jl")
+
+include("expr_abstraction.jl")
+export iscall
+export get_funsym
+export get_funarg
+export set_funarg
+
 include("EGraphs/EGraphs.jl")
 export EGraphs
 
