+++ test/runtests.jl
@@ -19,6 +19,8 @@ include("test_extraction.jl")
 include("test_mu.jl")
 include("test_while_interpreter.jl")
 include("test_symbolbug.jl")
+include("test_julog.jl")
+
 
 # include("test_while_superinterpreter.jl")
 
