+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"api/#API-Documentation","page":"API Documentation","title":"API Documentation","text":"","category":"section"},{"location":"api/#Metatheory","page":"API Documentation","title":"Metatheory","text":"","category":"section"},{"location":"api/","page":"API Documentation","title":"API Documentation","text":"Modules = [Metatheory]","category":"page"},{"location":"api/#Main.Metatheory.Theory","page":"API Documentation","title":"Main.Metatheory.Theory","text":"A Theory is either a vector of Rule or a compiled, callable function.\n\n\n\n\n\n","category":"type"},{"location":"api/#Main.Metatheory.Rule-Tuple{Expr}","page":"API Documentation","title":"Main.Metatheory.Rule","text":"Construct a Rule from a quoted expression. You can also use the [@rule] macro to create a Rule.\n\nSymbolic Rules\n\nRules defined as left_hand => right_hand are called symbolic rules. Application of a symbolic Rule is a replacement of the left_hand pattern with the right_hand substitution, with the correct instantiation of pattern variables. Function call symbols are not treated as pattern variables, all other identifiers are treated as pattern variables. Literals such as 5, :e, \"hello\" are not treated as pattern variables.\n\nDynamic Rules\n\nRules defined as left_hand |> right_hand are called dynamic rules. Dynamic rules behave like anonymous functions. Instead of a symbolic substitution, the right hand of a dynamic |> rule is evaluated during rewriting: matched values are bound to pattern variables as in a regular function call. This allows for dynamic computation of\n\nType Assertions\n\nType assertions are supported in the left hand of rules to match and access literal values both when using classic rewriting and EGraph based rewriting. To use a type assertion pattern, add ::T after a pattern variable in the left_hand of a rule.\n\n\n\nExamples\n\nSymbolic rule\n\nRule(:(a * b => b * a))\n\nDynamic rules computing the actual multiplication of two numbers on a match\n\nRule(:(a::Number * b::Number |> a*b))\n\n\n\nSignatures\n\nRule(e::Expr) -> Main.Metatheory.Rule\n\n\n\n\nMethods\n\nRule(e)\n\ndefined at /home/runner/work/Metatheory.jl/Metatheory.jl/src/rule.jl:65.\n\n\n\n\n\n","category":"method"},{"location":"api/#Main.Metatheory.compile_theory-Tuple{Array{Main.Metatheory.Rule,1},Module}","page":"API Documentation","title":"Main.Metatheory.compile_theory","text":"Compile a theory to a closure that does the pattern matching job Returns a RuntimeGeneratedFunction, which does not use eval and is as fast as a regular Julia anonymous function 🔥\n\n\n\nSignatures\n\ncompile_theory(theory::Array{Main.Metatheory.Rule,1}, mod::Module; __source__) -> RuntimeGeneratedFunctions.RuntimeGeneratedFunction{_A,_B,_C,_D} where _D where _C where _B where _A\n\n\n\n\nMethods\n\ncompile_theory(theory, mod; __source__)\n\ndefined at /home/runner/work/Metatheory.jl/Metatheory.jl/src/matchcore_compiler.jl:62.\n\n\n\n\n\n","categ
[…]
