### 2021-02-25 9e3ad37
minor optimizations and refactoring. working generalizing Expr interface
