### 2021-02-27 #3035 @0x0f0f0f
Extending the list of potential reviewers from the Google sheets file.

- dpsanders
- ChrisRackauckas
- vchuravy
- dmbates
- ahwillia
- ninjin
- ayushpatnaikgit
- dglmoore
- avik-pal
- pitsianis
- henrik227 (interested in symbolic computation)
- moorepants (interested in symbolic computation)
