### 2021-02-27 2661c59
Added `$` for interpolation in left hand of rules
