+++ src/Util/Util.jl
@@ -27,8 +27,9 @@ export amp
 
 include("cleaning.jl")
 export rmlines
-export binarize
+export binarize!
 export cleanast
+export interpolate_dollar
 
 include("walks.jl")
 export df_walk
@@ -39,6 +40,5 @@ export bf_walk!
 include("fixpoint.jl")
 export normalize
 export normalize_nocycle
-
 
 end
