### 2021-02-28 c07ab2a
fixed tests. some comments in README. compilation of dynamic rules RGF now happens at rule creation time
