### 2021-03-01 502465a
Fixed serious bug inconsistently deleting enodes from hashcons.
