+++ test/runtests.jl
@@ -20,8 +20,7 @@ include("test_julog.jl")
 include("test_boson.jl")
 include("test_theories.jl")
 include("test_reductions.jl")
-
-# include("test_while_superinterpreter.jl")
+include("test_while_superinterpreter.jl")
 
 # exported consistency test
 for m ∈ [Metatheory, Metatheory.Util, Metatheory.EGraphs, Metatheory.EGraphs.Schedulers]
