### 2021-03-02 41a831d
Removed type variables. Inconsistent and unnecessary (use abstract types)
