+++ src/EGraphs/ematch.jl
@@ -69,27 +69,30 @@ function ematch(e::EGraph, t::Expr, v::Int64, sub::Sub; lit=nothing)::Vector{Sub
     for n in e.M[find(e,v)]
         if isexpr(t, :(::))
             # right hand of type assertion
-            tr = t.args[2]
+            # tr = t.args[2]
+            typ = t.args[2]
 
-            if tr isa Type
-                typ = tr
-            elseif tr isa Symbol
-                if haskey(sub, tr)
-                    typ = sub[tr][2]
-                else
-                    # add the type to the egraph
-                    type_id = add!(e, typeof(n))
-                    sub =  Base.ImmutableDict(sub, t.args[2] => (type_id, typeof(n)))
-                    typ = typeof(n)
-                    # union!(c, ematch(e, t, v, sub))
-                    # continue
-                end
-            # elseif isexpr(tr, :curly)
-            # TODO allow for parametric type variables
-            # see https://dl.acm.org/doi/pdf/10.1145/3276483
-            else
-                error("Unsupported type assertion '", t, "'")
-            end
+            # println(n)
+            # println(typ, " ", typeof(typ))
+            # if tr isa Type
+            #     typ = tr
+            # elseif tr isa Symbol
+            #     if haskey(sub, tr)
+            #         typ = sub[tr][2]
+            #     else
+            #         # add the type to the egraph
+            #         type_id = add!(e, typeof(n))
+            #         sub =  Base.ImmutableDict(sub, t.args[2] => (type_id, typeof(n)))
+            #         typ = typeof(n)
+            #         # union!(c, ematch(e, t, v, sub))
+            #         # continue
+            #     end
+            # # elseif isexpr(tr, :curly)
+            # # TODO allow for parametric type variables
+            # # see https://dl.acm.org/doi/pdf/10.1145/3276483
+            # else
+            #     error("Unsupported type assertion '", t, "'")
+            # end
 
             !(typeof(n) <: typ) && continue
             union!(c, ematch(e, t.args[1], v, sub; lit=n))
