### 2021-03-02 #3078 @whedon
[REVIEW]: Metatheory.jl: Fast and Elegant Algebraic Computation in Julia with Extensible Equality Saturation
**Submitting author:** @ 0x0f0f0f (<a href="http://orcid.org/0000-0002-8122-9469">Alessandro Cheli</a>)
**Repository:** <a href="https://github.com/0x0f0f0f/Metatheory.jl" target ="_blank">https://github.com/0x0f0f0f/Metatheory.jl</a>
**Version:** v0.3.2

- [x] **Contribution and authorship:** Has the submitting author (@ 0x0f0f0f) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

- [x] **Contribution and authorship:** Has the submitting author (@ 0x0f0f0f) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

- [x] **Contribution and authorship:** Has the submitting author (@ 0x0f0f0f) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?
