### 2021-03-02 #3078 @0x0f0f0f
Thanks. I've tried to change the version number in the pre review but I did not have editor permissions to do so. Please keep in mind that a lot of things changed since v0.1.0 (many bugfixes) and the documentation in the repo is currently referring to the latest commit in the master branch of Metatheory.jl. The general concepts explained in the paper are unaltered.
