### 2021-03-02 #3078 @dpsanders
@0x0f0f0f: I think it would be helpful to include a simple example of the application of the library in the paper.
