### 2021-03-02 #3078 @dpsanders
@0x0f0f0f: The package will need proper narrative documentation, not just a quick README and the API reference.
See https://documentation.divio.com/ for a discussion of the different types of documentation required by a package.
