### 2021-03-03 #3078 @jpfairbanks
1. The main figure is a reproduction from the EGG paper. There should be a novel figure highlighting the contributions of this library over and above the EGG library.
2. The primary benefit of using Metatheory is the algebraic nature of the specification of a simplification system. I would like that to be highlighted in the text.
3. Examples of specifying algebraic theories should be provided including the composition of theories by concatenating their rules.
4. Why is concatenating rules sufficient for composition in this setting? Is it because these are all single-sorted theories with universally quantified equations?

I agree with David that this package needs additional narrative documentation. Some content in the readme would be appropriate for that documentation.
