### 2021-03-04 73ca272
Being lazy is now a property of an abstractanalysis (simplified the core egg code a lot)
