+++ test/test_extraction.jl
@@ -21,24 +21,9 @@ t = comm_monoid ∪ fold_mul
     extr = extract!(G, extran)
 	# println(extr)
 
-    @test extr == :((12*a) * b)
-end
-
-@testset "Lazy Extraction 1 - Commutative Monoid" begin
-    G = EGraph(:(3 * 4))
-    saturate!(G, t)
-	extran = addlazyanalysis!(G, ExtractionAnalysis, astsize)
-    @test (12 == extract!(G, extran))
-
-    ex = :(a * 3 * b * 4)
-    G = EGraph(cleanast(ex))
-	extran = addlazyanalysis!(G, ExtractionAnalysis, astsize)
-    saturate!(G, t)
-    extr = extract!(G, extran)
-	# println(extr)
-
     @test extr == :(b * (a * 12))
 end
+
 fold_add = @theory begin
 	a::$Number + b::$Number |> a + b
 end
@@ -69,7 +54,7 @@ end
     # sleep(0.3)
     ex = cleanast(:((x*(a+b)) + (y*(a+b))))
     G = EGraph(ex)
-	extran = addlazyanalysis!(G, ExtractionAnalysis, astsize)
+	extran = addanalysis!(G, ExtractionAnalysis, astsize)
     saturate!(G, t)
     # end
 
@@ -101,30 +86,6 @@ end
     @test extr == :((12 * a) * b)
 end
 
-@testset "Lazy Extraction - Adding analysis after saturation" begin
-    G = EGraph(:(3 * 4))
-    addexpr!(G, 12)
-    saturate!(G, t)
-    addexpr!(G, :(a * 2))
-    saturate!(G, t)
-
-    extran = addlazyanalysis!(G, ExtractionAnalysis, astsize)
-    saturate!(G, t)
-
-    @test (12 == extract!(G, extran))
-
-    # for i ∈ 1:100
-    ex = :(a * 3 * b * 4)
-    G = EGraph(cleanast(ex))
-    addanalysis!(G, NumberFold)
-    extran = addlazyanalysis!(G, ExtractionAnalysis, astsize)
-    saturate!(G, comm_monoid)
-
-    extr = extract!(G, extran)
-    # end
-
-    @test extr == :((12 * a) * b)
-end
 
 comm_monoid = @commutative_monoid (*) 1
 
@@ -178,35 +139,7 @@ t = comm_monoid ∪ comm_group ∪ distrib(:(*), :(+)) ∪ powers ∪ logids  �
 		saturate!(G, t; timeout=7)
 		ex = extract!(G, extran)
 	end
-	@test ex == :(log(a) * 5)
-end
-
-@testset "Complex Lazy Extraction" begin
-	G = EGraph(:(log(e) * log(e)))
-	extran = addlazyanalysis!(G, ExtractionAnalysis, astsize)
-	saturate!(G, t; timeout=7)
-	@test extract!(G, extran) == 1
-
-	G = EGraph(:(log(e) * (log(e) * e^(log(3)))  ))
-	extran = addlazyanalysis!(G, ExtractionAnalysis, astsize)
-	saturate!(G, t; timeout=7)
-	@test extract!(G, extran) == 3
-
-	@time begin
-		G = EGraph(:(a^3 * a^2))
-		extran = addlazyanalysis!(G, ExtractionAnalysis, astsize)
-		saturate!(G, t; timeout=7)
-		ex = extract!(G, extran)
-	end
-	@test ex == :(a^5)
-
-	@time begin
-		G = EGraph(:((log(e) * log(e)) * (log(a^3 * a^2))))
-		extran = addlazyanalysis!(G, ExtractionAnalysis, astsize)
[…]
