+++ src/EGraphs/EGraphs.jl
@@ -26,7 +26,6 @@ export AbstractAnalysis
 export merge!
 export addexpr!
 export addanalysis!
-export addlazyanalysis!
 export rebuild!
 
 include("analysis.jl")
