### 2021-03-04 #3078 @0x0f0f0f
Points

> 1. The main figure is a reproduction from the EGG paper. There should be a novel figure highlighting the contributions of this library over and above the EGG library.
> 2. The primary benefit of using Metatheory is the algebraic nature of the specification of a simplification system. I would like that to be highlighted in the text.
> 3. Examples of specifying algebraic theories should be provided including the composition of theories by concatenating their rules.
> 4. Why is concatenating rules sufficient for composition in this setting? Is it because these are all single-sorted theories with universally quantified equations?

These 4 points should have been added in the paper in the latest commit.
