+++ README.md
@@ -43,12 +43,8 @@ Or you can install the developer version (recommended by now for latest bugfixes
 ```julia
 julia> using Pkg; Pkg.add(url="https://github.com/0x0f0f0f/Metatheory.jl")
 ```
+## Usage
 
-## Please note that Metatheory.jl is in an experimental stage and THINGS ARE GOING TO CHANGE, A LOT
-
-## Examples
-
-Here are some of examples of the basic workflow of using Metatheory.jl. Theories are composable and reusable!
 Since Metatheory.jl relies on [RuntimeGeneratedFunctions.jl](https://github.com/SciML/RuntimeGeneratedFunctions.jl/), you have to call `@metatheory_init` in the module where you are going to use Metatheory.
 
 ```julia
@@ -58,253 +54,8 @@ using Metatheory.EGraphs
 @metatheory_init
 ```
 
-### Basic Symbolic Mathematics
+## Documentation
 
-#### Theories and Algebraic Structures
-The e-graphs backend can directly handle associativity,
-commutativity and distributivity, rules that are
-otherwise known of causing loops in symbolic computations.
+Narrative and API documentation for Metatheory.jl is available [here](https://0x0f0f0f.github.io/Metatheory.jl/dev)
 
-```julia
-comm_monoid = @theory begin
-    a * b => b * a
-    a * 1 => a
-    a * (b * c) => (a * b) * c
-end
-```
-
-#### The Metatheory Library
-
-
-The `Metatheory.Library` module contains utility functions and macros for creating
-rules and theories from commonly used algebraic structures and
-properties. This is equivalent to the previous theory definition.
-```julia
-using Metatheory.Library
-
-comm_monoid = commutative_monoid(:(*), 1)
-# alternatively
-comm_monoid = @commutative_monoid (*) 1
-```
-
-#### Theories are Collections and Composable
-
-Theories are just collections, precisely *vectors of the `Rule` object*, and can
-be composed as regular Julia collections. The most
-useful way of composing theories is unioning
-them with the '∪' operator.
-You are not limited to composing theories, you can
-manipulate and create them at both runtime and compile time
-as regular vectors.
-
-```julia
-comm_group = @theory begin
-    a + 0 => a
-    a + b => b + a
-    a + inv(a) => 0 # inverse
-    a + (b + c) => (a + b) + c
-end
-distrib = @theory begin
-    a * (b + c) => (a * b) + (a * c)
-end
-t = comm_monoid ∪ comm_group ∪ distrib
-```
-
-#### EGraphs
-
-An EGraph is an efficient data structure for representing congruence relations.
-Over the past decade, several projects have repurposed EGraphs to implement state-of-the-art, rewrite-driven compiler optimizations and program synthesizers using a technique known as equality saturation.
-Metatheory.jl provides a general purpose, customizable implementation of EGraphs and equality saturation, inspired from the [egg](https://egraphs-good.github.io/) library for Rust. You can read more about the design
-of the EGraph data structure and equality saturation algorithm in the
-[egg paper](https://dl.acm.org/doi/pdf/10.1145/3434304).
-
-#### What can I do with EGraphs in Metatheory.jl?
-
-Most importantly, the EGraph backend for Metatheory.jl allows
-you to create an EGraph from a starting expression, to add more expressions to the EGraph with `addexpr!`, and then to effectively fill the EGraph with all possible equivalent expressions resulting from applying rewrite rules from a theory, by using the `saturate!` function. You can then easily
[…]
