+++ test/runtests.jl
@@ -10,6 +10,7 @@ Metatheory.options[:verbose] = true
 
 @metatheory_init
 
+
 include("test_egraphs.jl")
 include("test_ematch.jl")
 include("test_ematch_assertions.jl")
@@ -24,6 +25,7 @@ include("test_theories.jl")
 include("test_reductions.jl")
 include("test_while_superinterpreter.jl")
 include("test_fibonacci.jl")
+include("test_cat.jl")
 
 # exported consistency test
 for m ∈ [Metatheory, Metatheory.Util, Metatheory.EGraphs, Metatheory.EGraphs.Schedulers]
