+++ docs/src/classic.md
@@ -1,0 +1,124 @@
+# Classical Rewriting
+
+There are some use cases where EGraphs and equality saturation are not
+required. The **classical rewriting backend** is suited for **simple tasks**
+when computing the whole equivalence class is overkill.
+
+The classical rewriting backend can be accessed with
+the [`rewrite`](@ref) function, which
+uses a recursive fixed point iteration algorithm to
+rewrite a source expression. The expression can be traversed with a
+**depth first** (inner left expression) evaluation order, or with a
+**breadth first** (outer left expression) evaluation order. You can
+configure the evaluation order by passing the keyword argument
+`order=:inner` (default) or `order=:outer` to the `rewrite` function.
+
+**Note**: With the classical [`rewrite`](@ref) algorithm, rules are matched in order and
+applied deterministically:
+every iteration, only the first rule that matches is applied.
+This means that when using the **classical rewriting backend**, the **ordering
+of rules in a theory matters!**. If some rules produce a loop, which is common
+for regular algebraic rules such as *commutativity, distributivity and associativity*,
+the other following rules in the theory *will never be applied*.
+
+The classical `rewrite`
+algorithm is suitable for:
+- Simple Pattern Matching Tasks
+- Interpretation of Code (e.g. interpretation of an eDSL)
+- Non-Optimizing Compiler Steps and Transformations (e.g. Your eDSL --> Julia)
+- Simple Deterministic Manipulation Tasks (e.g. cleaning expressions)
+
+For algebraic, mathematics oriented rewriting, please
+use the [`EGraph`](@ref) backend.
+
+Rewriting loops are detected by keeping an history of hashes of the
+rewritten expression. When a loop is detected, rewriting stops immediately
+and returns the current expression.
+
+Metatheory.jl is meant for
+composability: you can always compose and interleave rewriting steps that use
+the classical rewriting backend or the more advanced EGraph backend.
+
+## Example
+
+Let's simplify an expression in the `comm_monoid` theory
+by using the EGraph backend. After simplification,
+we may want to move all the `σ` symbols to the right of multiplications,
+we can do this simple task with a *classical rewriting* step, by using
+the `rewrite` function.
+
+##### Step 1: Simplification with EGraphs
+
+```jldoctest simpl
+using Metatheory
+using Metatheory.EGraphs
+using Metatheory.Library
+
+comm_monoid = commutative_monoid(:(*), 1);
+start_expr = :( (a * (1 * (2σ)) * (b * σ + (c * 1)) ) );
+g = EGraph(start_expr);
+extractor = addanalysis!(g, ExtractionAnalysis, astsize);
+saturate!(g, comm_monoid);
+simplified = extract!(g, extractor)
+
+# output
+:(a * (σ * 2) * (σ * b + c))
+
+```
+
+##### Step 2: Moving σ to the right
+```jldoctest simpl
+moveright = @theory begin
+	:σ * a 				=> a*:σ
+	(a * :σ) * b 	=> (a * b) * :σ
+	(:σ * a) * b 	=> (a * b) * :σ
+end;
+
+simplified = rewrite(simplified, moveright);
+
+# output
+:((a * (2 * :σ)) * (b * :σ + c))
+```
+
+#### Assignment to variables during rewriting.
+
+Using the *classical rewriting* backend, you may want
+to assign a value to an externally defined variable.
+Because of the nature of modules and the [`RuntimeGeneratedFunction`](https://github.com/SciML/RuntimeGeneratedFunctions.jl)
+compilation pipeline, it is not possible to assign
+values to variables in other modules.
[…]
