### 2021-03-04 51ec83c
add some tests for (monoidal) categories
