### 2021-03-05 3877d21
added equational rules with `==`
