+++ test/test_cat.jl
@@ -1,22 +1,22 @@
+using Metatheory
+using Metatheory.EGraphs
+using Test
 
 Cat = @theory begin
     1(A) => id(A).(A→A)
     id(A).(A→A) ⋅ f.(A→B) => f.(A→B)
-    f.(A→B) ⋅ id(A).(A→A) => f.(A→B)
     f.(A→B)⋅id(B).(B→B) => f.(A→B)
     (f.(A→B)⋅g.(B→C))⋅h.(C→D) => f.(A→B)⋅(g.(B→C)⋅h.(C→D))
 end
 
-using Test
 
 @testset "Cat" begin
-@test @areequal Cat 1(B) id(B).(B→B)
-@test @areequal Cat (id(A).(A→A)⋅f.(A→B)) f.(A→B)
-@test @areequal Cat (1(A)⋅f.(A→B)) f.(A→B)
-@test @areequal Cat (f.(A→B)⋅1(B)) f.(A→B)
-@test @areequal Cat (f.(A→B)⋅g.(B→B))⋅g.(B→B) f.(A→B)⋅(g.(B→B)⋅g.(B→B))
-@test @areequal Cat (f.(A→B)⋅g.(B→B))⋅h.(B→C) f.(A→B)⋅(g.(B→B)⋅h.(B→C))
-# areequal(CatTyped, :(f.(A→B)⋅1(B)), :f; timeout=100)
+    @test @areequal Cat 1(B) id(B).(B→B)
+    @test @areequal Cat (id(A).(A→A)⋅f.(A→B)) f.(A→B)
+    @test @areequal Cat (1(A)⋅f.(A→B)) f.(A→B)
+    @test @areequal Cat (f.(A→B)⋅1(B)) f.(A→B)
+    @test @areequal Cat (f.(A→B)⋅g.(B→B))⋅g.(B→B) f.(A→B)⋅(g.(B→B)⋅g.(B→B))
+    @test @areequal Cat (f.(A→B)⋅g.(B→B))⋅h.(B→C) f.(A→B)⋅(g.(B→B)⋅h.(B→C))
 end
 
 MonCat = @theory begin
@@ -32,6 +32,7 @@ MonCat = @theory begin
 
     (f⊗g)⊗h => f⊗(g⊗h)  # associativity for both obs and homs
     f.(A→B)⊗g.(C→D) => (f⊗g).(A⊗C→B⊗D)
+    (f⊗g).(A⊗C→B⊗D) => f.(A→B)⊗g.(C→D)
     id(A⊗B).(A⊗B→A⊗B) => id(A).(A→A)⊗id(B).(B→B)
 end
 
@@ -41,4 +42,28 @@ end
     @test @areequal MonCat 1(A⊗B)⋅(f.(A→B)⊗g.(B→C)) (f.(A→B)⊗g.(B→C))
     @test @areequal MonCat (f.(A→B)⊗g.(B→C))⋅ 1(A⊗B) (f.(A→B)⊗g.(B→C))
     @test @areequal MonCat 1(A⊗B)⋅(f.(A→B)⊗g.(B→C)) (f⊗g).(A⊗B→B⊗C)
+end
+
+SymMonCat = MonCat ∪ @theory begin
+    sigma(A, B) => σ(A, B).(A⊗B→B⊗A)
+    σ(A, B).(A⊗B→B⊗A) ⋅ σ(B,A).(B⊗A→A⊗B) => id(A⊗B).(A⊗B→A⊗B)
[…]
