### 2021-03-05 4580e9a
added equational rules with `==`
