### 2021-03-05 60c30f0
added mini calculational logic prover
