+++ test/runtests.jl
@@ -26,6 +26,8 @@ include("test_reductions.jl")
 include("test_while_superinterpreter.jl")
 include("test_fibonacci.jl")
 include("test_cat.jl")
+include("test_calculational_logic.jl")
+
 
 # exported consistency test
 for m ∈ [Metatheory, Metatheory.Util, Metatheory.EGraphs, Metatheory.EGraphs.Schedulers]
