### 2021-03-05 c9945bc
added mini calculational logic prover
