### 2021-03-05 31667ac
calculational logic improvements
