### 2021-03-05 2b04037
don't print iteration numbers by default
