### 2021-03-06 4578d23
issue 21 optimization
