+++ src/EGraphs/ematch.jl
@@ -101,8 +101,7 @@ function ematch(e::EGraph, t::Expr, v::Int64, sub::Sub; lit=nothing)::Vector{Sub
 
         # otherwise ematch on an expr
         (!(n isa Expr) || getfunsym(n) != getfunsym(t)) && continue
-        ids = getfunargs(n) .|> x -> x.id
-        ids = length(ids) == 0 ? Int64[] : ids
+        ids = Vector{Int64}(getfunargs(n) .|> x -> x.id)
         union!(c, ematchlist(e, getfunargs(t), ids, sub))
     end
     return c
