### 2021-03-06 2fc1ae5
tests, logging, elegant fix for issue 21
