### 2021-03-06 967328e
tests, logging, elegant fix for issue 21
