### 2021-03-06 e2f5c21
BUG: trying to implement context style type information
