### 2021-03-07 86e3f71
attempt at caching eclass ids where function symbols appear
