### 2021-03-07 9b26fd7
attempt at caching eclass ids where function symbols appear
