### 2021-03-07 5f7bf6d
refactoring saturation step
