+++ src/EGraphs/saturation.jl
@@ -40,27 +40,25 @@ function eqsat_step!(egraph::EGraph, theory::Vector{Rule};
         # outermost symbol in lhs
         sym = getfunsym(rule.left)
         ids = get(egraph.symcache, sym, [])
-        if rule.mode == :equational
-            symr = getfunsym(rule.right)
-            union!(ids, get(egraph.symcache, symr, []))
-        end
 
         for id ∈ ids
-        # for id ∈ get(egraph.symcache, sym, keys(egraph.M))
-        # for id ∈ keys(egraph.M)
-            # id = find(egraph, id)
-            # println(rule.right)
             for sub in ematch(egraph, rule.left, id, EMPTY_DICT)
                 # display(sub); println()
                 !isempty(sub) && push!(matches, (rule, sub, id))
             end
-            if rule.mode == :equational
+        end
+
+        if rule.mode == :equational
+            sym = getfunsym(rule.right)
+            ids = get(egraph.symcache, sym, [])
+            for id ∈ ids
                 for sub in ematch(egraph, rule.right, id, EMPTY_DICT)
                     # display(sub); println()
                     !isempty(sub) && push!(matches, (rule, sub, id))
                 end
             end
         end
+
 
     end
 
