### 2021-03-07 2b7ae8b
minimal improvements (avoid multiple dispatch where not needed)
