+++ test/test_calculational_logic.jl
@@ -73,7 +73,7 @@ t = calc ∪ fold
 @test @areequal t true ((p => (p ∨ p)) == ((¬(p) ∧ q) => q)) == true
 
 # Frege's theorem
-@test @areequal t true (p => (q => r)) => ((p => q) => (p => r))
+@test areequal(t, true, :((p => (q => r)) => ((p => q) => (p => r))); sizeout=2^14)
 
 # Demorgan's
 @test @areequal t true (¬(p ∨ q) == (¬p ∧ ¬q))
@@ -89,11 +89,11 @@ t = calc ∪ fold
 # @test @areequal (t ∪ [@rule :p => true]) true ((p => q) ∧ (r => s) ∧ (p ∨ r)) => (q ∨ s)
 
 #
-g = EGraph(:(((p => q) ∧ (r => s) ∧ (p ∨ r)) => (q ∨ s)))
-@time saturate!(g, t; timeout=10, sizeout=2^12)
-extran = addanalysis!(g, ExtractionAnalysis, astsize)
-
-println(extract!(g, extran))
+# g = EGraph(:(((p => q) ∧ (r => s) ∧ (p ∨ r)) => (q ∨ s)))
+# @time saturate!(g, t; timeout=10, sizeout=2^12)
+# extran = addanalysis!(g, ExtractionAnalysis, astsize)
+#
+# println(extract!(g, extran))
 
 #
 # in_same_set(g.U, g.root, addexpr!(g, true).id) |> println
