+++ test/test_analysis.jl
@@ -102,8 +102,8 @@ end
     ex = :(a * 3 * b * 4)
     G = EGraph(cleanast(ex))
     addanalysis!(G, NumberFold)
-    @test (true == areequal(G, comm_monoid, :((3 * a) * (4 * b)), :((12*a)*b),
-        :(((6*2)*b)*a)))
+    @test areequal(G, comm_monoid, :((3 * a) * (4 * b)), :((12*a)*b),
+        :(((6*2)*b)*a); timeout=15)
 end
 
 @testset "Infinite Loops analysis" begin
