+++ src/EGraphs/EGraphs.jl
@@ -35,6 +35,7 @@ include("analysis.jl")
 include("ematch.jl")
 include("Schedulers/Schedulers.jl")
 
+include("saturation_report.jl")
 include("saturation.jl")
 export saturate!
 include("equality.jl")
