+++ test/runtests.jl
@@ -6,7 +6,8 @@ using Metatheory.Util
 using DataStructures
 using Test
 
-Metatheory.options[:verbose] = true
+# Metatheory.options[:verbose] = true
+# Metatheory.options[:printiter] = true
 
 @metatheory_init
 
