+++ test/test_cat.jl
@@ -3,67 +3,82 @@ using Metatheory.EGraphs
 using Test
 
 Cat = @theory begin
+    id(A).(A→A) ⋅ f.(A→B) == f.(A→B)
+    f.(A→B)⋅id(B).(B→B) == f.(A→B)
+    (f.(A→B)⋅g.(B→C))⋅h.(C→D) == f.(A→B)⋅(g.(B→C)⋅h.(C→D))
+end
+
+# DOES NOT FIXPOINT!
+tag_matcher = @matcher begin
+    sigma(A, B) => σ(A, B).(A⊗B→B⊗A)
     1(A) => id(A).(A→A)
-    id(A).(A→A) ⋅ f.(A→B) => f.(A→B)
-    f.(A→B)⋅id(B).(B→B) => f.(A→B)
-    (f.(A→B)⋅g.(B→C))⋅h.(C→D) => f.(A→B)⋅(g.(B→C)⋅h.(C→D))
+    :f |> :(f.(A → B))
+    :g |> :(g.(B → C))
+    :h |> :(h.(B → C))
+    :j |> :(j.(B → B))
+    :k |> :(j.(C → D))
+end
+
+function tag(x)
+    r = df_walk(tag_matcher, x; skip_call=true, skip=[:(.)])
+    println("tagged $x to $(r)")
+    r
+end
+
+macro areequal_tag(theory, exprs...)
+    exprs = map(tag, exprs)
+
+    t = Metatheory.gettheory(theory, __module__; compile=false)
+    areequal(t, exprs...; mod=__module__)
 end
 
 
 @testset "Cat" begin
-    @test @areequal Cat 1(B) id(B).(B→B)
-    @test @areequal Cat (id(A).(A→A)⋅f.(A→B)) f.(A→B)
-    @test @areequal Cat (1(A)⋅f.(A→B)) f.(A→B)
-    @test @areequal Cat (f.(A→B)⋅1(B)) f.(A→B)
-    @test @areequal Cat (f.(A→B)⋅g.(B→B))⋅g.(B→B) f.(A→B)⋅(g.(B→B)⋅g.(B→B))
-    @test @areequal Cat (f.(A→B)⋅g.(B→B))⋅h.(B→C) f.(A→B)⋅(g.(B→B)⋅h.(B→C))
+    @test @areequal_tag Cat 1(B) id(B).(B→B)
+    @test @areequal_tag Cat (1(A) ⋅ f) f
+    @test @areequal_tag Cat (1(A)⋅f) f
+    @test @areequal_tag Cat (f⋅1(B)) f
+    @test @areequal_tag Cat (f⋅j)⋅j f⋅(j⋅j)
+    @test @areequal_tag Cat (f⋅j)⋅h f⋅(j⋅h)
 end
 
-MonCat = @theory begin
-    1(A) => id(A).(A→A)
-    id(A).(A→A) ⋅ f.(A→B) => f.(A→B)
-    f.(A→B) ⋅ id(A).(A→A) => f.(A→B)
-    f.(A→B)⋅id(B).(B→B) => f.(A→B)
-    (f.(A→B)⋅g.(B→C))⋅h.(C→D) => f.(A→B)⋅(g.(B→C)⋅h.(C→D))
-
-    munit() => munit()
-    A⊗munit() => A
[…]
