+++ src/Library/Library.jl
@@ -24,6 +24,7 @@ export identity_right
 export distrib_left
 export distrib_right
 export distrib
+export monoid
 export commutative_monoid
 export commutative_group
 export abelian_group
