### 2021-03-08 #3078 @mwillsey
Low-level feedback
- Line 26: "Theories can then be executed". Theories should be introduced first, this is an abrupt start.
- "rust" should be "Rust" throughout

Overall the paper, software, and documentation (especially) looks good. As the egg author, I view "egg in Julia" as a sufficient contribution, given that Metatheory.jl capitalizes on then dynamic nature of Julia with, e.g. homoiconicity, runtime code generation, and so on.

My only overall comment is that (classical, directed) rewriting deserves a slightly longer treatment. Lots of work has gone into making rewriting fast and effective (e.g. completion, reduction orders, etc.), and using equality saturation will be slower in many use cases; this is simply a tradeoff.
