### 2021-03-09 c44c785
Added egg generating script in benchmarks
