### 2021-03-09 #3078 @0x0f0f0f
Thank you max! I really appreciate your comments and suggestions. I wanted to keep this paper short and introductory (focusing on the important contributions) to leave a more theoretical (and analytical) approach to a follow-up. I'd be really happy to read and write more about classical rewriting. Please send me an email if you have any paper to suggest!
