### 2021-03-09 b657d4e
moved extraction analysis computation based on nodes children eclasses data to the cost fucntion. added taylor series test
