+++ test/runtests.jl
@@ -23,6 +23,7 @@ include("test_julog.jl")
 include("test_boson.jl")
 include("test_theories.jl")
 include("test_reductions.jl")
+include("test_taylor.jl")
 include("test_while_superinterpreter.jl")
 include("test_fibonacci.jl")
 include("test_cat.jl")
