### 2021-03-09 d1459be
moved extraction analysis computation based on nodes children eclasses data to the cost fucntion. added taylor series test
