+++ test/runtests.jl
@@ -6,8 +6,8 @@ using Metatheory.Util
 using DataStructures
 using Test
 
-# Metatheory.options[:verbose] = true
-# Metatheory.options[:printiter] = true
+Metatheory.options[:verbose] = true
+Metatheory.options[:printiter] = true
 
 @metatheory_init
 
@@ -17,7 +17,7 @@ include("test_ematch_assertions.jl")
 include("test_analysis.jl")
 include("test_dynamic_ematch.jl")
 include("test_extraction.jl")
-include("test_mu.jl")
+# FIXME include("test_mu.jl")
 include("test_while_interpreter.jl")
 include("test_julog.jl")
 include("test_boson.jl")
@@ -29,7 +29,7 @@ include("test_fibonacci.jl")
 include("test_cat.jl")
 include("test_calculational_logic.jl")
 include("test_logic.jl")
-# include("test_contexts.jl")
+include("test_contexts.jl")
 
 # exported consistency test
 for m ∈ [Metatheory, Metatheory.Util, Metatheory.EGraphs, Metatheory.EGraphs.Schedulers]
