+++ test/test_taylor.jl
@@ -15,6 +15,7 @@ x = big(42)
 
 b = rewrite(a, expand(5000)) |> eval
 # 1.739274941520501044994695988622883932193276720547806372656638132701531037200611e+18
+
 exp(x) + cos(x)
 # 1.739274941520501046994695988622883932193276720547806372656638132701531037200651e+18
 
