### 2021-03-10 b01ce34
Fixed cursed analysis bug
