+++ test/runtests.jl
@@ -29,11 +29,11 @@ include("test_fibonacci.jl")
 include("test_cat.jl")
 include("test_calculational_logic.jl")
 include("test_logic.jl")
-include("test_contexts.jl")
+# include("test_contexts.jl")
 
 # exported consistency test
 for m ∈ [Metatheory, Metatheory.Util, Metatheory.EGraphs, Metatheory.EGraphs.Schedulers]
    for i ∈ propertynames(m)
-      x = getproperty(m, i)
+      xxx = getproperty(m, i)
    end
 end
