+++ src/EGraphs/egg.jl
@@ -2,6 +2,7 @@
 # https://dl.acm.org/doi/10.1145/3434304
 
 using DataStructures
+using StaticArrays
 
 """
 Abstract type representing an [`EGraph`](@ref) analysis,
@@ -10,18 +11,15 @@ an EGraph
 """
 abstract type AbstractAnalysis end
 
-const ClassMem = Dict{Int64,Vector{Any}}
+
+const ClassMem = Dict{Int64,EClassData}
 const HashCons = Dict{Any,Int64}
-const Parent = Tuple{Any,Int64} # parent enodes and eclasses
 const ParentMem = Dict{Int64,Vector{Parent}}
 const AnalysisData = Dict{Int64,Any}
 const Analyses = Vector{AbstractAnalysis}
-
-"""
-a cache mapping function symbols to e-classes that
-contain e-nodes with that function symbol.
-"""
 const SymbolCache = Dict{Any, Vector{Int64}}
+
+
 
 """
 A concrete type representing an [`EGraph`].
@@ -34,12 +32,16 @@ mutable struct EGraph
     """map from eclass id to eclasses"""
     M::ClassMem             #
     H::HashCons             # hashcons
-    parents::ParentMem
+    # parents::ParentMem
     """worklist for ammortized upwards merging"""
     dirty::Vector{Int64}
     root::Int64
     """A vector of analyses associated to the EGraph"""
     analyses::Analyses
+    """
+    a cache mapping function symbols to e-classes that
+    contain e-nodes with that function symbol.
+    """
     symcache::SymbolCache
 end
 
@@ -47,7 +49,7 @@ EGraph() = EGraph(
     IntDisjointSets(0),
     ClassMem(),
     HashCons(),
-    ParentMem(),
+    # ParentMem(),
     Vector{Int64}(),
     0,
     Analyses(),
@@ -61,14 +63,6 @@ function EGraph(e)
     G
 end
 
-function addparent!(G::EGraph, a::Int64, parent::Parent)
-    @assert isenode(parent[1])
-    if !haskey(G.parents, a)
-        G.parents[a] = [parent]
-    else
-        union!(G.parents[a], [parent])
-    end
-end
 
 """
 Inserts an e-node in an [`EGraph`](@ref)
@@ -84,12 +78,17 @@ function add!(G::EGraph, n)::EClass
         return find(G, G.H[n]) |> EClass
     end
     @debug(n, " not found in H")
+
     id = push!(G.U) # create new singleton eclass
-    !haskey(G.parents, id) && (G.parents[id] = [])
-    getfunargs(n) .|> x -> addparent!(G, x.id, (n, id))
+
+    for child_eclass ∈ getfunargs(n)
+        addparent!(G.M[child_eclass.id], (n, id))
+    end
 
     G.H[n] = id
-    G.M[id] = [n]
+
+    classdata = EClassData(id, [n], [])
+    G.M[id] = classdata
 
     # cache the eclass for the symbol for faster matching
     sym = getfunsym(n)
@@ -119,14 +118,12 @@ function addexpr!(G::EGraph, e)::EClass
     df_walk((x -> add!(G, x)), e; skip_call = true)
 end
 
-function mergeparents!(G::EGraph, from::Int64, to::Int64)
-    !haskey(G.parents, from) && (G.parents[from] = []; return)
-    !haskey(G.parents, to) && (G.parents[to] = [])
-
-    union!(G.parents[to], G.parents[from])
-    delete!(G.parents, from)
+function clean_enode!(g::EGraph, t, to::Int64)
+    delete!(g.H, t)
+    canonicalize!(g.U, t)
[…]
