+++ test/test_analysis.jl
@@ -13,27 +13,28 @@ end
 NumberFold(g::EGraph) = NumberFold(g, Dict{Int64, Any}())
 
 # This should be auto-generated by a macro
-function EGraphs.make(an::NumberFold, n)
-    n isa Number && return n
+function EGraphs.make(an::NumberFold, n::ENode)
+    n.sym isa Number && (println("res = $n"); return n.sym)
 
-    if n isa Expr && Meta.isexpr(n, :call)
-        if n.args[1] == :*
-            l = n.args[2]
-            r = n.args[3]
+    println(n)
+    res = nothing
 
+    if ariety(n) > 0
+        l = n.args[1]
+        r = n.args[2]
+
+        if n.sym == :*
             if get(an, l, nothing) isa Number && get(an, r, nothing) isa Number
-                return an[l] * an[r]
+                res = an[l] * an[r]
             end
-        elseif n.args[1] == :+
-            l = n.args[2]
-            r = n.args[3]
-
+        elseif n.sym == :+
             if get(an, l, nothing) isa Number && get(an, r, nothing) isa Number
-                return an[l] + an[r]
+                res = an[l] + an[r]
             end
         end
     end
-    return nothing
+    println("res = ", res)
+    return res
 end
 
 function EGraphs.join(analysis::NumberFold, from, to)
@@ -50,7 +51,7 @@ function EGraphs.modify!(an::NumberFold, id::Int64)
     g = an.egraph
     # !haskey(an, id) && return nothing
     if an[id] isa Number
-        newclass = EGraphs.add!(g, an[id])
+        newclass = EGraphs.add!(g, ENode(an[id]))
         merge!(g, newclass.id, id)
     end
 end
