+++ src/EGraphs/extraction.jl
@@ -6,10 +6,8 @@ function astsize(n::ENode, an::AbstractAnalysis)
     cost = 1 + length(n.args)
 
     for a ∈ n.args
-        !haskey(an, a) && return Inf
-        if haskey(an, a) && an[a] != nothing
-            cost += last(an[a])
-        end
+        # !haskey(an, a) && return Inf
+        cost += last(an[a])
     end
     return cost
 end
@@ -21,7 +19,7 @@ Strives to get the largest expression
 """
 astsize_inv(n::ENode, an::AbstractAnalysis) = -1 * astsize(n, an)
 
-const CostData = Dict{Int64, Tuple{Any, Number}}
+const CostData = Dict{Int64, Tuple{ENode, Number}}
 
 """
 An [`AbstractAnalysis`](@ref) that computes the cost of expression nodes
@@ -52,11 +50,19 @@ Base.delete!(an::ExtractionAnalysis, id::Int64) = delete!(an.data, id)
 islazy(an::ExtractionAnalysis) = true
 
 function rec_extract(G::EGraph, an::ExtractionAnalysis, id::Int64)
+    println("extracting from", id)
+    println("class is ", G.M[id])
     (cn, ck) = an[id]
+    println("node is ", cn)
+    # cn = canonicalize(G.U, cn)
+    # println("canonicalized node is ", cn)
     (ariety(cn) == 0 || ck == Inf) && return cn.sym
 
     sym = cn.iscall ? :call : cn.sym
-    args = map(cn.args) do a rec_extract(G, an, a) end
+    args = map(cn.args) do a
+        # TODO evaluate this behaviour
+        rec_extract(G, an, a)
+    end
     args = cn.iscall ? [cn.sym, args...] : args
     return Expr(sym, args...)
 end
@@ -66,6 +72,7 @@ Given an [`ExtractionAnalysis`](@ref), extract the expression
 with the smallest computed cost from an [`EGraph`](@ref)
 """
 function extract!(G::EGraph, extran::ExtractionAnalysis)
+    println("root is $(G.root)")
     islazy(extran) && analyze!(G, extran, G.root)
     !(extran ∈ G.analyses) && error("Extraction analysis is not associated to EGraph")
     rec_extract(G, extran, G.root)
