### 2021-03-11 0e4824c
improvements and invariant assertions
