+++ test/test_analysis.jl
@@ -14,9 +14,8 @@ NumberFold(g::EGraph) = NumberFold(g, Dict{Int64, Any}())
 
 # This should be auto-generated by a macro
 function EGraphs.make(an::NumberFold, n::ENode)
-    n.sym isa Number && (println("res = $n"); return n.sym)
+    n.sym isa Number && return n.sym
 
-    println(n)
     res = nothing
 
     if ariety(n) > 0
@@ -93,9 +92,9 @@ end
     addanalysis!(G, NumberFold)
     saturate!(G, comm_monoid)
 
-    display(G.M); println()
+    # display(G.M); println()
     println(G.root)
-    display(G.analyses[1].data); println()
+    # display(G.analyses[1].data); println()
 
     @test (true == areequal(G, comm_monoid, :(3 * 4), 12, :(4*3), :(6*2)))
 
