+++ test/runtests.jl
@@ -6,8 +6,8 @@ using Metatheory.Util
 using DataStructures
 using Test
 
-Metatheory.options[:verbose] = true
-Metatheory.options[:printiter] = true
+# Metatheory.options[:verbose] = true
+# Metatheory.options[:printiter] = true
 
 @metatheory_init
 
@@ -15,26 +15,26 @@ Test.FallbackTestSet(desc) = Test.FallbackTestSet()
 
 ts = Test.FallbackTestSet
 
-@testset ts "Metatheory Tests" begin
-   @timev begin
+# @testset ts "Metatheory Tests" begin
+   # @timev begin
       include("test_egraphs.jl")
       include("test_ematch.jl")
       include("test_ematch_assertions.jl")
       include("test_analysis.jl")
-      include("test_extraction.jl")
-      include("test_dynamic_ematch.jl")
-      # FIXME include("test_mu.jl")
-      include("test_while_interpreter.jl")
-      include("test_julog.jl")
-      include("test_boson.jl")
-      include("test_theories.jl")
-      include("test_reductions.jl")
-      include("test_taylor.jl")
-      include("test_while_superinterpreter.jl")
-      include("test_fibonacci.jl")
-      include("test_cat.jl")
-      include("test_calculational_logic.jl")
-      include("test_logic.jl")
+      # include("test_extraction.jl")
+      # include("test_dynamic_ematch.jl")
+      # # FIXME include("test_mu.jl")
+      # include("test_while_interpreter.jl")
+      # include("test_julog.jl")
+      # include("test_boson.jl")
+      # include("test_theories.jl")
+      # include("test_reductions.jl")
+      # include("test_taylor.jl")
+      # include("test_while_superinterpreter.jl")
+      # include("test_fibonacci.jl")
+      # include("test_cat.jl")
+      # include("test_calculational_logic.jl")
+      # include("test_logic.jl")
       # include("test_contexts.jl")
       # exported consistency test
       for m ∈ [Metatheory, Metatheory.Util, Metatheory.EGraphs, Metatheory.EGraphs.Schedulers]
@@ -42,5 +42,5 @@ ts = Test.FallbackTestSet
             xxx = getproperty(m, i)
          end
       end
-   end
-end
+   # end
+# end
