### 2021-03-11 1c0337d
improvements and invariant assertions
