+++ src/EGraphs/analysis.jl
@@ -21,9 +21,6 @@ end
 """
 function analyze!(g::EGraph, analysis::AbstractAnalysis, ids::Vector{Int64})
     ids = sort(ids)
-
-    println(ids)
-
     @assert isempty(g.dirty)
 
     did_something = true
@@ -51,7 +48,7 @@ function analyze!(g::EGraph, analysis::AbstractAnalysis, ids::Vector{Int64})
     for id ∈ ids
         id = find(g, id)
         if !haskey(analysis, id)
-            display(g.M[id]); println()
+            # display(g.M[id]); println()
             # display(analysis.data); println()
             error("failed to compute analysis for eclass ", id)
         end
