+++ docs/Project.toml
@@ -1,4 +1,5 @@
 [deps]
+AutoHashEquals = "15f4f7f2-30c1-5605-9d31-71845cf9641f"
 Calculus = "49dc2e85-a5d0-5ad3-a950-438e2897f1b9"
 DataStructures = "864edb3b-99cc-5e75-8d2d-829cb0a9cfe8"
 DocStringExtensions = "ffbed154-4ef7-542d-bbb7-c09d3a79fcae"
@@ -7,3 +8,4 @@ DocumenterTools = "35a29f4d-8980-5a13-9543-d66fff28ecb8"
 MatchCore = "5dd3f0b1-72a9-48ad-ae6e-79f673da005f"
 Metatheory = "e9d8d322-4543-424a-9be4-0cc815abe26c"
 RuntimeGeneratedFunctions = "7e49a35a-f44a-4d26-94aa-eba1b4ca6b47"
+StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
