### 2021-03-11 a3e23f7
release bump and benchmarks
