+++ Project.toml
@@ -1,7 +1,7 @@
 name = "Metatheory"
 uuid = "e9d8d322-4543-424a-9be4-0cc815abe26c"
 authors = ["Alessandro Cheli - 0x0f0f0f <sudo-woodo3@protonmail.com>"]
-version = "0.1.1"
+version = "0.2"
 
 [deps]
 AutoHashEquals = "15f4f7f2-30c1-5605-9d31-71845cf9641f"
