+++ benchmarks/Project.toml
@@ -1,0 +1,6 @@
+[deps]
+BenchmarkTools = "6e4b80f9-dd63-53aa-95a3-0cdb28fa8baf"
+Metatheory = "e9d8d322-4543-424a-9be4-0cc815abe26c"
+Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
+Suppressor = "fd094767-a336-5f1f-9728-57cf17d0bbfb"
+SymbolicUtils = "d1185830-fcd6-423d-90d6-eec64667417b"
