### 2021-03-11 f286f06
Added guarded category theory file
