+++ test/test_cat_guard.jl
@@ -1,0 +1,200 @@
+using Metatheory
+using Metatheory.EGraphs
+using Metatheory.Util
+# Description here:
+# https://www.philipzucker.com/metatheory-progress/
+
+
+# https://github.com/AlgebraicJulia/Catlab.jl/blob/ce2fde9c63a8aab65cf2a7697f43cd24e5e00b3a/src/theories/Monoidal.jl#L127
+
+cat_rules = @theory begin
+    f ⋅ id(b) => f
+    id(a) ⋅ f => f
+
+    #f => f ⋅ id(dom(type(f)))
+    #f => id(cod(type(f))) ⋅ f
+
+    a ⊗ₒ munit() => a
+    munit() ⊗ₒ a => a
+
+    #a => a ⊗ₒ munit() 
+    #a => munit() ⊗ₒ a
+
+    f ⋅ (g ⋅ h) == (f ⋅ g) ⋅ h
+
+
+end
+
+monoidal_rules = @theory begin
+    id(munit()) ⊗ₘ f => f
+    f ⊗ₘ id(munit()) => f
+    a ⊗ₒ (b ⊗ₒ c) == (a ⊗ₒ b) ⊗ₒ c
+    f ⊗ₘ (h ⊗ₘ j) == (f ⊗ₘ h) ⊗ₘ j
+    id(a ⊗ₒ b) == id(a) ⊗ₘ id(b)
+
+    (f ⊗ₘ p) ⋅ (g ⊗ₘ q) |>
+    # future metatheory macros will clean this up
+    begin
+        fcod = addexpr!(_egraph, :(cod(type($f))))
+        gdom = addexpr!(_egraph, :(dom(type($g))))
+        pcod = addexpr!(_egraph, :(cod(type($p))))
+        qdom = addexpr!(_egraph, :(dom(type($q))))
+        if (fcod == gdom && pcod == qdom)
+            :(($f ⋅ $g) ⊗ₘ ($p ⋅ $q))
+        else
+            :(($f ⊗ₘ $p) ⋅ ($g ⊗ₘ $q))
+        end
+    end
+
+    (f ⋅ g) ⊗ₘ (p ⋅ q) => (f ⊗ₘ p) ⋅ (g ⊗ₘ q)
+end
+
+
+
+sym_rules = @theory begin
+    σ(a, b) ⋅ σ(b, a) == id(a ⊗ₒ b)
+    (σ(a, b) ⊗ₘ id(c)) ⋅ (id(b) ⊗ₘ σ(a, c)) == σ(a, b ⊗ₒ c)
+    (id(a) ⊗ₘ σ(b, c)) ⋅ (σ(a, c) ⊗ₘ id(b)) == σ(a ⊗ₒ b, c)
+
+    (f ⊗ₘ h) ⋅ σ(a, b) |> begin
+        fcod = addexpr!(_egraph, :(cod(type($f)))).id
+        hcod = addexpr!(_egraph, :(cod(type($h)))).id
+        if fcod == a && hcod == b
+            :(σ(dom(type($f)), dom(type($h))) ⋅ ($h ⊗ₘ $f))
+        else
+            :(($f ⊗ₘ $h) ⋅ σ($a, $b))
+        end
+    end
+
+
+    σ(c, d) ⋅ (h ⊗ₘ f) |> begin
[…]
