### 2021-03-11 e453f5d
_lhs_expr optimization!
