+++ test/test_julog.jl
@@ -13,6 +13,7 @@ end
 JulogAnalysis(g::EGraph, facts::Vector{Clause}) =
     JulogAnalysis(g, facts, Dict{Int64, Vector{Clause}}())
 
+# TODO adapt analysis to new ENode implementation
 # This should be auto-generated by a macro
 function EGraphs.make(an::JulogAnalysis, n)
     !(n isa Expr) && return []
@@ -57,7 +58,7 @@ macro res(goals, facts)
 end
 
 example = @theory begin
-    x/x |> (@res(not(iszero(x)), facts) ? :($x) : :($x/$x))
+    x/x |> (@res(not(iszero(x)), facts) ? :($x) : _lhs_expr)
 end
 
 display(g.M); println()
