+++ src/Util/cleaning.jl
@@ -31,3 +31,6 @@ interp_dol(any, mod::Module) = any
 function interpolate_dollar(ex, mod::Module)
     df_walk(interp_dol, ex, mod)
 end
+
+
+remove_assertions(e) = df_walk( x -> (isexpr(x, :(::)) ? x.args[1] : x), e; skip_call=true )
