### 2021-03-11 033f5b0
_lhs_expr optimization!
