+++ Project.toml
@@ -1,7 +1,7 @@
 name = "Metatheory"
 uuid = "e9d8d322-4543-424a-9be4-0cc815abe26c"
 authors = ["Alessandro Cheli - 0x0f0f0f <sudo-woodo3@protonmail.com>"]
-version = "0.2"
+version = "0.2.0"
 
 [deps]
 AutoHashEquals = "15f4f7f2-30c1-5605-9d31-71845cf9641f"
@@ -12,6 +12,9 @@ RuntimeGeneratedFunctions = "7e49a35a-f44a-4d26-94aa-eba1b4ca6b47"
 StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 
 [compat]
+AutoHashEquals = "0.2"
+DocStringExtensions = "0.8"
+StaticArrays = "1"
 DataStructures = "0.18.9"
 MatchCore = "0.1.0"
 RuntimeGeneratedFunctions = "0.5.1"
