### 2021-03-12 2f8331c
algorithm simplification. improvements to enode structure
