+++ test/test_egraphs.jl
@@ -1,13 +1,6 @@
 
 # ENV["JULIA_DEBUG"] = Metatheory
 
-@testset "e-node validation" begin
-    @test EGraphs.isenode(2) == true
-    @test EGraphs.isenode( :(2 + 3)  ) == false
-    @test EGraphs.isenode( EClass(2) ) == false
-    @test EGraphs.isenode( Expr(:call, :foo, EClass(2)) ) == true
-    @test EGraphs.isenode( Expr(:call, :foo, 3, EClass(2)) ) == false
-end
 
 ## SPAZZATURA
 
@@ -37,7 +30,6 @@ end
     @test in_same_set(G.U, t2.id, t1.id)
     println(find_root!(G.U, t2.id))
     @test find_root!(G.U, t2.id) == 4
-    # display(G.parents); println()
     rebuild!(G)
     # f(a,b) = f(a,c)
     # display(G.M); println()
