+++ test/test_analysis.jl
@@ -14,7 +14,7 @@ NumberFold(g::EGraph) = NumberFold(g, Dict{Int64, Any}())
 
 # This should be auto-generated by a macro
 function EGraphs.make(an::NumberFold, n::ENode)
-    n.sym isa Number && return n.sym
+    n.head isa Number && return n.head
 
     res = nothing
 
@@ -22,11 +22,11 @@ function EGraphs.make(an::NumberFold, n::ENode)
         l = n.args[1]
         r = n.args[2]
 
-        if n.sym == :*
+        if n.head == :*
             if get(an, l, nothing) isa Number && get(an, r, nothing) isa Number
                 res = an[l] * an[r]
             end
-        elseif n.sym == :+
+        elseif n.head == :+
             if get(an, l, nothing) isa Number && get(an, r, nothing) isa Number
                 res = an[l] + an[r]
             end
