### 2021-03-12 959bf42
algorithm simplification. improvements to enode structure
