### 2021-03-12 59b000e
term interface abstraction
