### 2021-03-12 5f82e76
term interface abstraction
