+++ src/Metatheory.jl
@@ -23,16 +23,10 @@ end
 
 export options
 
-
-
 # TODO document this interface
-include("expr_abstraction.jl")
-export iscall
-export istree
-export getfunsym
-export getfunargs
-export setfunsym!
-export setfunargs!
+include("TermInterface.jl")
+using .TermInterface
+export TermInterface
 
 include("Util/Util.jl")
 using .Util
