+++ src/Util/Util.jl
@@ -3,12 +3,7 @@ module Util
 
 include("../docstrings.jl")
 
-import ..iscall
-import ..getfunsym
-import ..getfunargs
-import ..setfunsym!
-import ..setfunargs!
-
+using ..TermInterface
 
 using Base.Meta
 ## AST manipulation utility functions
