### 2021-03-12 ad47acb
updated addexpr for new interface
