### 2021-03-12 396fc8d
updated addexpr for new interface
