### 2021-03-12 4422601
updated library with bidir rules
