+++ src/Library/algebra.jl
@@ -1,8 +1,8 @@
-commutativity(op) = :( ($op)(a, b) => ($op)(b, a) ) |> Rule
+commutativity(op) = :( ($op)(a, b) == ($op)(b, a) ) |> Rule
 right_associative(op) = :( ($op)(a, $(op)(b,c)) => ($op)($(op)(a,b), c) ) |> Rule
 left_associative(op) = :( ($op)($(op)(a,b), c) => ($op)(a, $(op)(b,c)) ) |> Rule
 
-associativity(op) = [right_associative(op), left_associative(op)]
+associativity(op) = :( ($op)(a, $(op)(b,c)) == ($op)($(op)(a,b), c) ) |> Rule
 
 identity_left(op, id) = :( ($op)($id, a) => a ) |> Rule
 identity_right(op, id) = :( ($op)(a, $id) => a ) |> Rule
@@ -10,16 +10,30 @@ identity_right(op, id) = :( ($op)(a, $id) => a ) |> Rule
 inverse_left(op, id, invop) = :( ($op)(($invop)(a), a) => $id ) |> Rule
 inverse_right(op, id, invop) = :( ($op)(a, ($invop)(a)) => $id ) |> Rule
 
+# distributivity of two operations
+# example: `@distrib (⋅) (⊕)`
+function distrib_left(outop, inop)
+	@assert Base.isbinaryoperator(outop)
+	@assert Base.isbinaryoperator(inop)
+	:( ($outop)(a, ($inop)(b,c)) == ($inop)(($outop)(a,b),($outop)(a,c)) ) |> Rule
+end
+
+function distrib_right(outop, inop)
+	@assert Base.isbinaryoperator(outop)
+	@assert Base.isbinaryoperator(inop)
+	:( ($outop)(($inop)(a,b), c) == ($inop)(($outop)(a,c),($outop)(b,c)) ) |> Rule
+end
+
 function monoid(op, id)
 	@assert Base.isbinaryoperator(op)
-	[right_associative(op), left_associative(op), identity_left(op, id), identity_right(op,id)]
+	[associativity(op), identity_left(op, id), identity_right(op,id)]
 end
 macro monoid(op, id) monoid(op, id) end
 
 
 function commutative_monoid(op, id)
 	@assert Base.isbinaryoperator(op)
-	[commutativity(op), right_associative(op), identity_left(op, id)]
+	[commutativity(op), associativity(op), identity_left(op, id)]
 end
 macro commutative_monoid(op, id) commutative_monoid(op, id) end
 
@@ -36,34 +50,9 @@ abelian_group(op, id, invop) = commutative_group(op, id, invop)
 macro commutative_group(op, id, invop) commutative_group(op, id, invop) end
 macro abelian_group(op, id, invop) commutative_group(op, id, invop) end
 
-# distributivity of two operations
-# example: `@distrib (⋅) (⊕)`
-function distrib_left(outop, inop)
-	@assert Base.isbinaryoperator(outop)
-	@assert Base.isbinaryoperator(inop)
-	:( ($outop)(a, ($inop)(b,c)) => ($inop)(($outop)(a,b),($outop)(a,c)) ) |> Rule
-end
 
-function distrib_right(outop, inop)
-	@assert Base.isbinaryoperator(outop)
-	@assert Base.isbinaryoperator(inop)
-	:( ($outop)(($inop)(a,b), c) => ($inop)(($outop)(a,c),($outop)(b,c)) ) |> Rule
-end
-
-function rev_distrib_left(outop, inop)
-	@assert Base.isbinaryoperator(outop)
-	@assert Base.isbinaryoperator(inop)
[…]
