### 2021-03-14 c1fdf43
big performance improvements and refactoring
