### 2021-03-14 7b7c424
big performance improvements and refactoring
