+++ src/EGraphs/saturation.jl
@@ -8,6 +8,8 @@ import ..@log
 
 using .Schedulers
 
+using Memoize
+
 function inst_step(pat, sub::Sub, side::Symbol)
     # TODO interface istree (?)
     if haskey(sub, pat)
@@ -25,6 +27,7 @@ function inst_step(pat, sub::Sub, side::Symbol)
     end
 end
 
+# @memoize ??? 
 function inst(pat, sub::Sub, side::Symbol)
     # remove type assertions
     if side == :left
@@ -36,41 +39,66 @@ function inst(pat, sub::Sub, side::Symbol)
     f
 end
 
+"""
+Exactly like [`addexpr!`](@ref), but instantiate pattern variables
+from a substitution, resulting from a pattern matcher run.
+"""
+function addexprinst_rec!(G::EGraph, pat, sub::Sub, side::Symbol)::EClass
+    # e = preprocess(pat)
+    # println("========== $e ===========")
+    class_ids = Int64[]
 
-"""
-Core algorithm of the library: the equality saturation step.
-"""
-function eqsat_step!(egraph::EGraph, theory::Vector{Rule};
-        scheduler=SimpleScheduler(), mod=@__MODULE__,
-        match_hist=MatchesBuf(), sizeout=options[:sizeout], stopwhen=()->false,
-        matchlimit=options[:matchlimit] # max number of matches
-        )
+    if haskey(sub, pat)
+        (eclass, lit) = sub[pat]
+        pat = (lit != nothing ? lit : eclass)
+    elseif pat isa Symbol
+        error("unbound pattern variable $pat")
+    elseif side == :right && pat isa QuoteNode && pat.value isa Symbol
+        pat = pat.value
+    end
+
+    for child ∈ getargs(pat)
+        c_eclass = addexprinst!(G, child, sub, side)
+        push!(class_ids, c_eclass.id)
+    end
+
+    node = ENode(pat, class_ids)
+
+    return add!(G, node)
+end
+
+addexprinst_rec!(g::EGraph, e::EClass, sub::Sub)::EClass = e
+
+addexprinst!(g::EGraph, e, sub::Sub, side::Symbol)::EClass =
+    addexprinst_rec!(g, preprocess(e), sub, side)
+
+# @memoize
+function sataddexpr!(g::EGraph, e)
+    addexpr!(g,e)
+end
+
+function cached_ids(egraph::EGraph, side)::Vector{Int64}
+    # outermost symbol in rule side
+    if istree(side)
+        sym = gethead(side)
+        get(egraph.symcache, sym, [])
+    else
+        collect(keys(egraph.M))
+    end
+end
+
+function eqsat_search!(egraph::EGraph, theory::Vector{Rule},
+        scheduler::AbstractScheduler)::MatchesBuf
     matches=MatchesBuf()
-
-    report = Report(egraph)
-    instcache = Dict{Rule, Dict{Sub, Int64}}()
-
-
-    readstep!(scheduler)
-
-
-    search_stats = @timed for rule ∈ theory
+    for rule ∈ theory
         # don't apply banned rules
-        # !haskey(match_hist, rule) && (match_hist[rule] = Set{Sub}())
-        # !haskey(matches, rule) && (matches[rule] = Set{Sub}())
         shouldskip(scheduler, rule) && continue
 
         if rule.mode ∉ [:rewrite, :dynamic, :equational]
             error("unsupported mode in rule ", rule)
         end
 
-        # outermost symbol in lhs
-        sym = gethead(rule.left)
-        if istree(rule.left)
-            ids = get(egraph.symcache, sym, [])
-        else
-            ids = keys(egraph.M)
-        end
+        ids = cached_ids(egraph, rule.left)
 
[…]
