### 2021-03-14 ab01780
fix memoize
