+++ src/Classic/Classic.jl
@@ -1,0 +1,28 @@
+"""
+This module contains classical rewriting functions and utilities
+"""
+module Classic
+
+import ..Rule
+import ..Theory
+# import ..gettheory
+import ..closure_generator
+import ..@metatheory_init
+using ..Util
+using Base.Meta
+
+@metatheory_init
+
+include("matchcore_compiler.jl")
+include("rewrite.jl")
+include("match.jl")
+
+export rewrite
+export @rewrite
+export @esc_rewrite
+export @compile_theory
+export @matcher
+export @rewriter
+
+
+end
