+++ src/EGraphs/enode.jl
@@ -1,6 +1,5 @@
 using DataStructures
 using Base.Meta
-using StaticArrays
 using AutoHashEquals
 
 import Base.ImmutableDict
@@ -12,16 +11,16 @@ end
 
 @auto_hash_equals struct ENode{X}
     head::Any
-    args::MVector{T, Int64} where T
+    args::Vector{Int64}
     metadata::Union{Nothing, NamedTuple}
 end
 
 ariety(n::ENode) = length(n.args)
 
 function ENode(e, c_ids::AbstractVector{Int64})
-    @assert length(getargs(e)) == length(c_ids)
-    static_args = MVector{length(c_ids), Int64}(c_ids...)
-    ENode{typeof(e)}(gethead(e), static_args, getmetadata(e))
+    # @assert length(getargs(e)) == length(c_ids)
+    # static_args = MVector{length(c_ids), Int64}(c_ids...)
+    ENode{typeof(e)}(gethead(e), c_ids, getmetadata(e))
 end
 
 ENode(e) = ENode(e, Int64[])

[…]
