### 2021-03-15 e3406fa
huge speedups. refactoring of classical backend in custom module. added some docs for rule syntax
