+++ .github/FUNDING.yml
@@ -1,6 +1,6 @@
 # These are supported funding model platforms
 
-github: [0x0f0f0f] # Replace with up to 4 GitHub Sponsors-enabled usernames e.g., [user1, user2]
+# github: [0x0f0f0f] # Replace with up to 4 GitHub Sponsors-enabled usernames e.g., [user1, user2]
 #patreon: # Replace with a single Patreon username
 #open_collective: # Replace with a single Open Collective username
 #ko_fi: # Replace with a single Ko-fi username
@@ -10,3 +10,4 @@ github: [0x0f0f0f] # Replace with up to 4 GitHub Sponsors-enabled usernames e.g.
 #issuehunt: # Replace with a single IssueHunt username
 #otechie: # Replace with a single Otechie username
 #custom: # Replace with up to 4 custom sponsorship URLs e.g., ['link1', 'link2']
+custom: ['https://github.com/sponsors/0x0f0f0f/']
