+++ Project.toml
@@ -1,7 +1,7 @@
 name = "Metatheory"
 uuid = "e9d8d322-4543-424a-9be4-0cc815abe26c"
 authors = ["Alessandro Cheli - 0x0f0f0f <sudo-woodo3@protonmail.com>"]
-version = "0.2.0"
+version = "0.2.1"
 
 [deps]
 AutoHashEquals = "15f4f7f2-30c1-5605-9d31-71845cf9641f"
@@ -9,7 +9,6 @@ DataStructures = "864edb3b-99cc-5e75-8d2d-829cb0a9cfe8"
 DocStringExtensions = "ffbed154-4ef7-542d-bbb7-c09d3a79fcae"
 MatchCore = "5dd3f0b1-72a9-48ad-ae6e-79f673da005f"
 RuntimeGeneratedFunctions = "7e49a35a-f44a-4d26-94aa-eba1b4ca6b47"
-StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 
 [compat]
 AutoHashEquals = "0.2"
@@ -17,7 +16,6 @@ DataStructures = "0.18.9"
 DocStringExtensions = "0.8"
 MatchCore = "0.1.0"
 RuntimeGeneratedFunctions = "0.5.1"
-StaticArrays = "1"
 julia = "1"
 
 [extras]
