### 2021-03-15 #3078 @0x0f0f0f
@jpfairbanks how does the current documentation look? Anything you think I should add?
Thank you Max! Very interesting paper. Well written and full of examples.
