### 2021-03-15 713800f
added anti-rules, minor optimizations
