+++ src/EGraphs/saturation.jl
@@ -8,40 +8,6 @@ import ..@log
 
 using .Schedulers
 
-# using Memoize
-
-function inst_step(pat, sub::Sub, side::Symbol)
-    if pat isa Expr
-        start = isexpr(pat, :call) ? 2 : 1
-        pat.args[start:end] = pat.args[start:end] .|> x ->
-            inst_step(x, sub, side)
-            return pat
-    end
-
-    if haskey(sub, pat)
-        (eclass, lit) = sub[pat]
-        lit != nothing ? lit : eclass
-    elseif pat isa Symbol
-        error("unbound pattern variable $pat")
-    elseif side == :right && pat isa QuoteNode && pat.value isa Symbol
-        pat.value
-    else
-        pat
-    end
-end
-
-# @memoize ???
-function inst(pat, sub::Sub, side::Symbol)
-    # remove type assertions
-    if side == :left
-        pat = remove_assertions(pat)
-    end
-
-    expr = deepcopy(pat)
-    inst_step(expr, sub, side)
-    # df_walk(inst_step, pat, sub, side; skip_call=true)
-end
-
 """
 Exactly like [`addexpr!`](@ref), but instantiate pattern variables
 from a substitution, resulting from a pattern matcher run.
@@ -79,17 +45,11 @@ function addexprinst_rec!(G::EGraph, pat, sub::Sub, side::Symbol)::EClass
     end
 
     return add!(G, ENode(pat))
-
-    # println("node $node")
-
 end
-
-# addexprinst_rec!(g::EGraph, e::EClass, sub::Sub, side::Symbol)::EClass = (println("matched eclass $e"); e)
 
 addexprinst!(g::EGraph, e, sub::Sub, side::Symbol)::EClass =
     addexprinst_rec!(g, preprocess(e), sub, side)
 
-# @memoize
 
 function cached_ids(egraph::EGraph, side)::Vector{Int64}
     # outermost symbol in rule side
@@ -108,7 +68,7 @@ function eqsat_search!(egraph::EGraph, theory::Vector{Rule},
         # don't apply banned rules
         shouldskip(scheduler, rule) && continue
 
-        if rule.mode ∉ [:rewrite, :dynamic, :equational]
+        if rule.mode ∉ [:rewrite, :dynamic, :equational, :inequality]
             error("unsupported mode in rule ", rule)
         end
 
@@ -121,7 +81,7 @@ function eqsat_search!(egraph::EGraph, theory::Vector{Rule},
             end
         end
 
-        if rule.mode == :equational
+        if rule.mode == :equational || rule.mode == :inequality
             ids = cached_ids(egraph, rule.right)
             for id ∈ ids
                 for sub in ematch(egraph, rule.right, id)
@@ -157,39 +117,19 @@ function eqsat_apply!(egraph::EGraph, matches::MatchesBuf,
 
         writestep!(scheduler, rule)
 
-        if rule.mode == :rewrite || rule.mode == :equational # symbolic replacement
-            # l = inst(rule.left, sub, :left)
-            # r = inst(rule.right, sub, :right)
-            l = remove_assertions(rule.left)
+        l = remove_assertions(rule.left)
+        lc = addexprinst!(egraph, l, sub, :left)
+        if rule.mode == :rewrite || rule.mode == :equational || rule.mode == :inequality # symbolic replacement
             r = rule.right
-            # r = unquote_sym(rule.right)
-            # println(l); println(r)
-            lc = addexprinst!(egraph, l, sub, :left)
             rc = addexprinst!(egraph, r, sub, :right)
-            # lc = addexpr!(egraph, l)
-            # rc = addexpr!(egraph, r)
-            merge!(egraph, lc.id, rc.id)
-
[…]
