### 2021-03-15 1391587
added anti-rules, minor optimizations
