### 2021-03-16 3fda922
some refactoring. various optimizations in saturation
