+++ src/Classic/matchcore_compiler.jl
@@ -38,7 +38,7 @@ function compile_rule(rule::Rule)::Expr
 	if rule.mode == :equational
 		error("equational rules not yet supported by classic rewriting backend." *
 			"Knuth-Bendix completion algorithm has not yet been implemented.")
-	elseif rule.mode == :inequality
+	elseif rule.mode == :unequal
 		error("anti-rules not yet supported by classical backend")
     elseif rule.mode == :dynamic # regular pattern matching
         # right side not quoted! needed to evaluate expressions in right hand.
@@ -47,7 +47,7 @@ function compile_rule(rule::Rule)::Expr
 			_lhs_expr = $(Meta.quot(ll));
 			$(rule.right)
 		end
-    elseif rule.mode == :rewrite
+    elseif rule.mode == :symbolic
 		# right side is quoted, symbolic replacement
         re = df_walk(c_right, rule.right, patvars; skip=skips, skip_call=true) |> quot
 	else
