+++ test/test_logic.jl
@@ -109,14 +109,22 @@ end
 ex = rewrite(:(((p => q) ∧ (r => s) ∧ (p ∨ r)) => (q ∨ s)), impl)
 @test prove(t, ex, 2)
 
+# using Metatheory.EGraphs.Schedulers
+# # println(ex)
 # ex = rewrite(:(((p => q) ∧ (r => s) ∧ (p ∨ r)) => (q ∨ s)), impl)
-# println(ex)
 # g = EGraph(ex)
-# @profiler saturate!(g, t; timeout=8, sizeout=2^15)
-
+# @timev repo = saturate!(g, t; timeout=10, sizeout=2^15, scheduler=ScoredScheduler)
+# println(repo)
+#
 # extran = addanalysis!(g, ExtractionAnalysis, astsize)
 # ex = extract!(g, extran)
 # println(ex)
+# #
+# #
+# ex = rewrite(:(((p => q) ∧ (r => s) ∧ (p ∨ r)) => (q ∨ s)), impl)
+# g = EGraph(ex)
+# @timev areequal(g, t, ex, true; timeout=10, sizeout=2^15, scheduler=ScoredScheduler, schedulerparams=(32, 1))
+#
 
 # @profiler saturate!(g, t; timeout=8, sizeout=2^15)
 # exit(0)
