+++ test/test_inequality.jl
@@ -7,5 +7,6 @@ end
 
 g = EGraph(:foo)
 report = saturate!(g, failme)
+println(report)
 @test report.reason == :contradiction
 @test !(@areequal failme foo wazoo)

[…]
