### 2021-03-16 21611e3
some refactoring. various optimizations in saturation
