+++ test/runtests.jl
@@ -25,9 +25,9 @@ ts = Test.FallbackTestSet
       include("test_extraction.jl")
       include("test_dynamic_ematch.jl")
       # FIXME include("test_mu.jl")
-      include("test_while_interpreter.jl")
       include("test_julog.jl")
       include("test_boson.jl")
+      include("test_while_interpreter.jl")
       include("test_theories.jl")
       include("test_reductions.jl")
       include("test_taylor.jl")
