+++ src/EGraphs/Schedulers/Schedulers.jl
@@ -4,6 +4,7 @@ include("../../docstrings.jl")
 
 import ..Rule
 import ..EGraph
+using ..Util
 
 include("./abstractscheduler.jl")
 include("./backoffscheduler.jl")
