### 2021-03-16 #3078 @philzook58
I have already found Metatheory.jl to be an extremely useful addition to the space with thoughtful insights in regards to usability, flexibility, and integration with the host language.

I have an intermixing of syntactic/grammar/style comments and actual content questions. I marked these comments as Style, since their suggestions are even more optional. I hope I am not overstepping my purview as a reviewer.

- " a well-known paradigm similar to several LISP idioms such as Scheme." >> I wouldn't call Scheme a "Lisp idiom". I would call it a dialect. Suggestion : "a well-known paradigm found in LISP dialects such as Scheme".  I've had discussions about what exactly homoiconicity is and I'm personally not sure. I'm unclear if it is a "paradigm" persay. But this is very subjective. Wikipedia seems to use the term "property of a programming language" which feels more appropriate to me. https://en.wikipedia.org/wiki/Homoiconicity

- Style - "Julia programming language, designed to take advantage of >> its << powerful reflection capabilities to"

- "Intuitively, Metatheory.jl transforms Julia expressions in other Julia expressions at both compile and run time.". I think this makes sense and I see what your going for, but Julia has a looser notion of compile and run time than most languages. I don't have a good suggestion.

- Style - "Theories can then be executed throught two, highly composable, rewriting backends. The first one, is based just on standard rewriting built on top of the pattern matcher developed in [@matchcore]." I think you do not need any of these commas. There is also a typo "throught"

- Style - "Such an approach suffers of the usual problems of rewriting systems. Even trivial equational rules such as commutativity may lead to non-terminating systems and thus need to be adjusted by some sort of structuring or rewriting order. This requires extensive user reasoning."

- Style - "The other back-end for Metatheory.jl, the core of our contribution, is designed to not require the user to reason about rewriting order by relying on >> equality saturation on e-graphs, the state-of-the-art technique adapted from the egg rust library [@egg]. <<" adjustment of sentence.

-  Provided with a theory of equational rewriting rules, defined in pure Julia, e-graphs compactly represent many equivalent programs.  >>> An egraph can be usefully used without equational pattern matching rewriting rules by merely inserting a fixed number of ground terms and equivalences into it. This is the usual mode in which they are used in quantifier free SMT queries. I feel this may be conflating two distinct aspects of the egraph technique as inseparable.

- "Saturation iteratively executes an e-graph specific pattern matcher to efficiently compute (and analyze) all possible equivalent expressions contained in the e-graph congruence closure." >>> In my understanding, the definition of the egraph includes that it propagates all congruence closures at some point. The term saturation refers to saturation with respect to the rules, not with respect to the congruence closure.

- Style - " beyond being easy to use and composable, " >> I erased "to be"

- Style - I would suggest removing basically all colons ":" throughout the paper and just starting new sentences.

-  "such as costs of nodes or logical statements attached to terms." The logical statements part is unclear to me how it would work. I think this may be referring to discussions we've had so I want to bring this up. Is this an expected thing to be able to do or something already done?

- "the most fitting expressions from an >> equivalence class represented in an e-graph." >> I don't think congruence closure is the appropriate term here.

- Style - "Many applications of equality saturation have been recently published, tailoring advanced optimization tasks." This use of the word "tailoring" feels a little odd to me. Suggestion: "Many applications of equality saturation to advanced optimization tasks have been recently published."

- Style - "In [@yang2021equality], authors used egg to superoptimize tensor signal flow graphs describing neural networks. However, Herbie requires interoperation and conversion of expressions between different languages and libraries." I would switch these sentences. Finish talking about Herbie.
