### 2021-03-16 #3078 @philzook58
I think I need a new invitation link as I can't edit my checkboxes and it appears to have expired. My apologies

I would also like to note for the sake of transparency that I have been in private discussion with the author on progress and usage of the library before even this article was submitted (it is mentioned in the acknowledgements of the paper also). We have no other prior connections besides that. I do not consider this to be a true conflict of interest.
