### 2021-03-16 #3078 @0x0f0f0f
@philzook58 I don't think it is a true conflict of interest. AFAIK you've been constantly reviewing my implementation decisions even before I've submitted the paper.
