+++ src/EGraphs/eclass.jl
@@ -3,8 +3,8 @@
 const Parent = Tuple{ENode,Int64} # parent enodes and eclasses
 mutable struct EClassData
     id::Int64
-    nodes::Vector{ENode}
-    parents::Vector{Parent}
+    nodes::Set{ENode}
+    parents::Dict{ENode, Int64}
 end
 
 # Interface for indexing EClassData
@@ -20,20 +20,18 @@ Base.iterate(a::EClassData, state) = iterate(a.nodes, state)
 # Showing
 Base.show(io::IO, a::EClassData) = Base.show(io, a.nodes)
 
-EClassData(id::Int64) = EClassData(id, [], Parent[])
+EClassData(id::Int64) = EClassData(id, Set{ENode}(), Dict{ENode, Int64}())
 
-function addparent!(a::EClassData, parent::Parent)
-    if parent ∉ a.parents
-        push!(a.parents, parent)
-    end
+function addparent!(a::EClassData, n::ENode, id::Int64)
+    a.parents[n] = id
 end
 
-function Base.union(from::EClassData, to::EClassData)
+function Base.union(to::EClassData, from::EClassData)
     EClassData(to.id, from.nodes ∪ to.nodes, from.parents ∪ to.parents)
 end
 
 function Base.union!(to::EClassData, from::EClassData)
     union!(to.nodes, from.nodes)
-    union!(to.parents, from.parents)
+    merge!(to.parents, from.parents)
     return to
 end
