### 2021-03-16 8184c93
paper fixes from Phil and Max
