### 2021-03-16 #3078 @0x0f0f0f
Updated the pdf with corrections from Phil
