+++ test/test_extraction.jl
@@ -23,8 +23,9 @@ t = comm_monoid ∪ fold_mul
     extr = extract!(G, extran)
 	println(extr)
 
-    @test extr == :(b * (a * 12)) || extr == :((b * 12) * a) || extr == :(a * (b * 12)) ||
-		extr == :((a * b) * 12) || extr == :((12a) * b) || extr == :(a * (12b))
+	@test extr == :((12 * a) * b) || extr == :(12 * (a * b)) || extr == :(a * (b * 12)) ||
+		extr == :((a * b) * 12) || extr == :((12a) * b) || extr == :(a * (12b)) ||
+		extr == :((b * (12a))) || extr == :((b * 12) * a)
 end
 
 fold_add = @theory begin
@@ -88,7 +89,7 @@ end
 
     @test extr == :((12 * a) * b) || extr == :(12 * (a * b)) || extr == :(a * (b * 12)) ||
 		extr == :((a * b) * 12) || extr == :((12a) * b) || extr == :(a * (12b)) ||
-		extr == :((b * (12a)))
+		extr == :((b * (12a))) || extr == :((b * 12) * a)
 end
 
 
