### 2021-03-16 beac44a
Fixed different results on same call
