### 2021-03-17 #3078 @jpfairbanks
I agree with @philzook58's suggestions. I think that the Metatheory library is rapidly evolving as we review this document, and I am not sure about the editorial standard for JOSS. Are libraries supposed to be mostly mature at the time of JOSS publication? Metatheory.jl is rapidly gaining new features, which might prompt a delay in JOSS publication, but I would leave that decision to the editors including @dpsanders. For example, since the original submission, Metatheory now defines an interface for term like types, which enables support for user supplied types to be used for terms. This is a large improvement over the original EGG implementation and should be called out in this article. I'm not sure what the editorial standard on this is. I am happy to see improvements in the library, but I think we will need a new JOSS submission in a month to document all the new features. Then again, if JOSS entries are living documents that always point to the latest version of the libraries docs, then this is less of a concern.

I also have been in contact with @0x0f0f0f about the integration of typed expressions with the goal of using Metatheory in my own work.
