### 2021-03-17 cc16ae6
even more speed, moved parameters to their saturation type
