### 2021-03-17 ca8ee5c
even more speed, moved parameters to their saturation type
