### 2021-03-17 #3078 @0x0f0f0f
I'm planning to write another full length paper in detail explaining all the features! This is an introductory paper to serve as canonical citation in the style of `Flux.jl`. In this paper I want to illustrate the core features and high level architecture of Metatheory.jl (egraphs and classical rewriting), which are unlikely to change.
