### 2021-03-17 #3078 @0x0f0f0f
I made a single breaking change and I've updated the example in the paper. Will be present in next commit pushed.
