+++ test/test_boson.jl
@@ -10,13 +10,11 @@ function boson_expand(ex)
     params = SaturationParams(timeout=4)
     G = EGraph(ex)
     saturate!(G, boson ∪ distr_assoc, params)
-    extractor = addanalysis!(G, ExtractionAnalysis, astsize_inv)
-    ex = extract!(G, extractor)
+    ex = extract!(G, astsize_inv)
 
     G = EGraph(ex)
     saturate!(G, distr_assoc ∪ ident, params)
-    extractor = addanalysis!(G, ExtractionAnalysis, astsize)
-    ex = extract!(G, extractor)
+    ex = extract!(G, astsize)
     # ex = rewrite(ex, ident)
     println(ex)
     return ex
