+++ src/EGraphs/egg.jl
@@ -8,10 +8,9 @@ Abstract type representing an [`EGraph`](@ref) analysis,
 attaching values from a join semi-lattice domain to
 an EGraph
 """
-abstract type AbstractAnalysis end
 const ClassMem = Dict{Int64,EClassData}
 const HashCons = Dict{ENode,Int64}
-const Analyses = Vector{AbstractAnalysis}
+const Analyses = Set{Type{<:AbstractAnalysis}}
 const SymbolCache = Dict{Any, Vector{Int64}}
 
 
@@ -50,11 +49,12 @@ EGraph() = EGraph(
     SymbolCache()
 )
 
+
 function EGraph(e)
-    G = EGraph()
-    rootclass = addexpr!(G, e)
-    G.root = rootclass.id
-    G
+    g = EGraph()
+    rootclass = addexpr!(g, e)
+    g.root = rootclass.id
+    g
 end
 
 function canonicalize(g::EGraph, n::ENode)
@@ -75,8 +75,12 @@ end
 """
 Returns the canonical e-class id for a given e-class.
 """
-find(G::EGraph, a::Int64)::Int64 = find_root!(G.U, a)
-find(G::EGraph, a::EClass)::Int64 = find_root!(G.U, a.id)
+find(g::EGraph, a::Int64)::Int64 = find_root!(g.U, a)
+find(g::EGraph, a::EClass)::Int64 = find_root!(g.U, a.id)
+
+
+geteclass(g::EGraph, a::Int64)::EClassData = g.M[find(g, a)]
+geteclass(g::EGraph, a::EClass)::Int64 = geteclass(g, a.id)
 
 
 ### Definition 2.3: canonicalization
@@ -88,38 +92,38 @@ iscanonical(g::EGraph, e::EClass) = find(g, e.id) == e.id
 """
 Inserts an e-node in an [`EGraph`](@ref)
 """
-function add!(G::EGraph, n::ENode)::EClass
+function add!(g::EGraph, n::ENode)::EClass
     @debug("adding ", n)
 
-    n = canonicalize!(G, n)
-    if haskey(G.H, n)
-        return find(G, G.H[n]) |> EClass
+    n = canonicalize!(g, n)
+    if haskey(g.H, n)
+        return find(g, g.H[n]) |> EClass
     end
     @debug(n, " not found in H")
 
-    id = push!(G.U) # create new singleton eclass
+    id = push!(g.U) # create new singleton eclass
 
     for c_id ∈ n.args
-        addparent!(G.M[c_id], n, id)
+        addparent!(g.M[c_id], n, id)
     end
 
-    G.H[n] = id
+    g.H[n] = id
 
     classdata = EClassData(id, OrderedSet([n]), OrderedDict{ENode, Int64}())
-    G.M[id] = classdata
+    g.M[id] = classdata
 
     # cache the eclass for the symbol for faster matching
     sym = n.head
-    if !haskey(G.symcache, sym)
-        G.symcache[sym] = Int64[]
+    if !haskey(g.symcache, sym)
+        g.symcache[sym] = Int64[]
     end
-    push!(G.symcache[sym], id)
+    push!(g.symcache[sym], id)
 
     # make analyses for new enode
-    for analysis ∈ G.analyses
-        if !islazy(analysis)
-            analysis[id] = make(analysis, n)
-            modify!(analysis, id)
+    for an ∈ g.analyses
+        if !islazy(an)
+            setdata!(classdata, an, make(an, g, n))
+            modify!(an, g, id)
         end
     end
 
@@ -131,7 +135,7 @@ Recursively traverse an [`Expr`](@ref) and insert terms into an
[…]
