### 2021-03-17 60185d0
completely redesigned the analysis interface to be based on abstract types only! easier to define and to work with!
