+++ test/test_analysis.jl
@@ -4,38 +4,31 @@ using MatchCore
 
 # ENV["JULIA_DEBUG"] = Metatheory
 
-struct NumberFold <: AbstractAnalysis
-    egraph::EGraph
-    data::Dict{Int64, Any}
-end
-
-# Mandatory for AbstractAnalysis
-NumberFold(g::EGraph) = NumberFold(g, Dict{Int64, Any}())
+abstract type NumberFold <: AbstractAnalysis end
 
 # This should be auto-generated by a macro
-function EGraphs.make(an::NumberFold, n::ENode)
+function EGraphs.make(an::Type{NumberFold}, g::EGraph, n::ENode)
     n.head isa Number && return n.head
 
-    res = nothing
+    if ariety(n) == 2
+        l = g.M[n.args[1]]
+        r = g.M[n.args[2]]
+        ldata = getdata(l, an, nothing)
+        rdata = getdata(r, an, nothing)
 
-    if ariety(n) > 0
-        l = n.args[1]
-        r = n.args[2]
-
-        if n.head == :*
-            if get(an, l, nothing) isa Number && get(an, r, nothing) isa Number
-                res = an[l] * an[r]
-            end
-        elseif n.head == :+
-            if get(an, l, nothing) isa Number && get(an, r, nothing) isa Number
-                res = an[l] + an[r]
+        if ldata isa Number && rdata isa Number
+            if n.head == :*
+                return ldata * rdata
+            elseif n.head == :+
+                return ldata + rdata
             end
         end
     end
-    return res
+
+    return nothing
 end
 
-function EGraphs.join(analysis::NumberFold, from, to)
+function EGraphs.join(an::Type{NumberFold}, from, to)
     if from isa Number
         if to isa Number
             @assert from == to
@@ -45,19 +38,17 @@ function EGraphs.join(analysis::NumberFold, from, to)
     return to
 end
 
-function EGraphs.modify!(an::NumberFold, id::Int64)
-    g = an.egraph
+function EGraphs.modify!(an::Type{NumberFold}, g::EGraph, id::Int64)
     # !haskey(an, id) && return nothing
-    if an[id] isa Number
-        newclass = EGraphs.add!(g, ENode(an[id]))
+    eclass = g.M[id]
+    d = getdata(eclass, an, nothing)
+    if d isa Number
+        newclass = addexpr!(g, d)
         merge!(g, newclass.id, id)
     end
 end
 
-Base.setindex!(an::NumberFold, value, id::Int64) = setindex!(an.data, value, id)
-Base.getindex(an::NumberFold, id::Int64) = an.data[id]
-Base.haskey(an::NumberFold, id::Int64) = haskey(an.data, id)
-Base.delete!(an::NumberFold, id::Int64) = delete!(an.data, id)
+EGraphs.islazy(x::Type{NumberFold}) = false
 
 comm_monoid = @theory begin
     a * b => b * a
@@ -66,7 +57,10 @@ comm_monoid = @theory begin
 end
 
 G = EGraph(:(3 * 4))
-addanalysis!(G, NumberFold)
+analyze!(G, NumberFold)
+
+# exit(0)
+
 @testset "Basic Constant Folding Example - Commutative Monoid" begin
     # addanalysis!(G, NumberFold())
     @test (true == @areequalg G comm_monoid 3 * 4 12)
@@ -79,8 +73,6 @@ end
     ex = :(a * 3 * b * 4)
     G = EGraph(cleanast(ex))
     addanalysis!(G, NumberFold)
-    display(G.M); println()
-    display(G.analyses[1].data); println()
[…]
