### 2021-03-17 #3078 @0x0f0f0f
Only change is that on the last line of example you do not need anymore to add the analysis explicitly but you can directly call `extract!(egraph, cost_function)`. @philzook58 everything looking ok? Were you able to edit the whedon comment with the list after Daniel re-added you?
