### 2021-03-17 2b7001e
unified interface for EClass (no shallow reference). added CAS test example with typeanalysis
