### 2021-03-17 dfecea5
unified interface for EClass (no shallow reference). added CAS test example with typeanalysis
