+++ docs/Project.toml
@@ -7,5 +7,6 @@ Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
 DocumenterTools = "35a29f4d-8980-5a13-9543-d66fff28ecb8"
 MatchCore = "5dd3f0b1-72a9-48ad-ae6e-79f673da005f"
 Metatheory = "e9d8d322-4543-424a-9be4-0cc815abe26c"
+Parameters = "d96e819e-fc66-5662-9728-84c9c7592b0a"
 RuntimeGeneratedFunctions = "7e49a35a-f44a-4d26-94aa-eba1b4ca6b47"
 StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
