+++ README.md
@@ -48,14 +48,3 @@ Narrative and API documentation for Metatheory.jl is available [here](https://0x
 ## Citing
 
 If you use Metatheory.jl in your research, please [cite](https://github.com/0x0f0f0f/Metatheory.jl/blob/master/CITATION.bib) our works.
-
-```
-@misc{cheli2021metatheoryjl,
-      title={Metatheory.jl: Fast and Elegant Algebraic Computation in Julia with Extensible Equality Saturation},
-      author={Alessandro Cheli},
-      year={2021},
-      eprint={2102.07888},
-      archivePrefix={arXiv},
-      primaryClass={cs.PL}
-}
-```
