+++ test/test_cas.jl
@@ -3,10 +3,10 @@ using Metatheory.Library
 using Metatheory.EGraphs
 using Metatheory.Classic
 using Metatheory.Util
-using EGraphs.Schedulers
+using Metatheory.EGraphs.Schedulers
 
 ## TypeAnalysis
-
+@metatheory_init
 # M is a Module
 abstract type TypeAnalysis <: AbstractAnalysis end
 
@@ -23,7 +23,7 @@ function EGraphs.make(an::Type{TypeAnalysis}, g::EGraph, n::ENode{T}) where T
             # unknown type for CAS
             t = Any
         end
-        println("analyzed type of $n is $t")
+        # println("analyzed type of $n is $t")
         return t
     end
 
@@ -83,15 +83,91 @@ function infer(e)
     getdata(geteclass(g, g.root), TypeAnalysis)
 end
 
-mult_t = commutative_monoid(:(*), 1)
-plus_t = commutative_monoid(:(+), 0)
-
-theory = mult_t ∪ plus_t
 
 ex1 = :(cos(1 + 3.0) + 4 + (4-4im))
 ex2 = :("ciao" * 2)
 ex3 = :("ciao" * " mondo")
 
-infer(ex1)
-infer(ex2)
-infer(ex3)
+@test ComplexF64 == infer(ex1)
+@test_throws ErrorException infer(ex2)
+@test String == infer(ex3)
+
+## Theory for CAS
+
+
+mult_t = commutative_monoid(:(*), 1)
+plus_t = commutative_monoid(:(+), 0)
+
+minus_t = @theory begin
+    a - a       => 0
+    a + (-b)    => a - b
+end
+
+mulplus_t = @theory begin
+    a * 0       => 0
+    a * (b + c) == ((a*b) + (a*c))
+    a + (b * a) => ((b+1)*a)
+end
+
+pow_t = @theory begin
+    (y^n) * y   => y^(n+1)
+    x^n * x^m   == x^(n+m)
+    (x * y)^z   == x^z * y^z
+    (x^p)^q     == x^(p*q)
+    x^0         => 1
+    x^1         => x
+    inv(x)      == x^(-1)
+end
+
+# Dynamic rules
+fold_t = @theory begin
+    -(a::Number)            |> -a
+    a::Number + b::Number   |> a + b
+    a::Number * b::Number   |> a * b
+end
+
+cas = fold_t ∪ mult_t ∪ plus_t ∪ minus_t ∪ mulplus_t #∪ pow_t
+
+using Metatheory.TermInterface
+
+function customlt(x,y)
+    if typeof(x) == Expr && Expr == typeof(y)
+        false
+    elseif typeof(x) == typeof(y)
+        isless(x,y)
+    elseif x isa Symbol && y isa Number
+        false
+    else
+        true
+    end
+end
+
+canonical_t = @theory begin
+    # restore n-arity
+    (x + (+)(ys...)) => +(x,ys...)
+    ((+)(xs...) + y) => +(xs..., y)
+    (x * (*)(ys...)) => *(x,ys...)
+    ((*)(xs...) * y) => *(xs..., y)
+
+    (*)(xs...)      |> Expr(:call, :*, sort!(xs; lt=customlt)...)
+    (+)(xs...)      |> Expr(:call, :+, sort!(xs; lt=customlt)...)
+end
+
+macro simplify(ex)
+    g = EGraph(ex)
+    println(saturate!(g, cas))
+    res = extract!(g, astsize)
+    println(res)
+    # for (id, ec) ∈ g.M
+    #     println(id, " => ", collect(ec.nodes))
+    #     println("\t\t", getdata(ec, ExtractionAnalysis{astsize}))
[…]
