+++ test/test_ematch_assertions.jl
@@ -7,10 +7,9 @@
     end
 
     G = EGraph(:(2*3))
-    display(G.M); println()
+    
 
     @test true == areequal(G, some_theory, :(2 * 3), :(matched(2,3)))
-    display(G.M); println()
     @test true == areequal(G, some_theory, :(matched(2,3)), :(specific(3,2)))
 end
 
