+++ test/test_fibonacci.jl
@@ -12,8 +12,6 @@ g = EGraph(:(fib(10)))
 z = EGraph(:(fib(10)))
 @time saturate!(z, fibo, params)
 
-display(g.M); println()
-
 @testset "Fibonacci" begin
     @test 55 == extract!(g, astsize)
 end
