### 2021-03-19 a4f0cab
fixed pattern matcher bug #40 and other minor improvements
