+++ test/test_cas.jl
@@ -8,7 +8,9 @@ using Metatheory.EGraphs.Schedulers
 
 ## TypeAnalysis
 @metatheory_init
-# M is a Module
+
+
+## M is a Module
 abstract type TypeAnalysis <: AbstractAnalysis end
 
 # This should be auto-generated by a macro
@@ -104,7 +106,7 @@ minus_t = @theory begin
     a - b       => a + (-1*b)
     -a          => -1 * a
     # TODO FIXME this rules bugs the pattern matcher
-    # a + (-b)    => a + (-1*b)
+    a + (-b)    => a + (-1*b)
 end
 
 mulplus_t = @theory begin
@@ -145,7 +147,8 @@ fold_t = @theory begin
     a::Number / b::Number   |> a/b
 end
 
-cas = fold_t ∪ mult_t ∪ plus_t ∪ minus_t ∪ mulplus_t ∪ pow_t ∪ div_t
+# cas = fold_t ∪ mult_t ∪ plus_t ∪ minus_t ∪ mulplus_t ∪ pow_t ∪ div_t
+cas = fold_t ∪ mult_t ∪ plus_t ∪ minus_t ∪ mulplus_t
 
 using Metatheory.TermInterface
 
@@ -178,7 +181,7 @@ end
 Metatheory.options[:verbose] = false
 Metatheory.options[:printiter] = false
 
-macro simplify(ex)
+function simplify(ex)
     rep = @timev begin
         g = EGraph(ex)
         # simpl_hist = []
@@ -201,41 +204,62 @@ macro simplify(ex)
         #     println(id, " => ", collect(ec.nodes))
         #     println("\t\t", getdata(ec, ExtractionAnalysis{astsize}))
         # end
-        res = rewrite(res, canonical_t; clean=false, m=__module__)
-        Meta.quot(res)
+        res = rewrite(res, canonical_t; clean=false, m=@__MODULE__)
     end
     rep
 end
 
+macro simplify(ex)
+    Meta.quot(simplify(ex))
+end
 
 
-@simplify 2a + a + a
-@simplify a * c * b
-@simplify 1 * x * 2
-@simplify (a*b)^2
+@test :(4a) == @simplify 2a + a + a
+@test :(a*b*c) @simplify a * c * b
+@test :(3x) @simplify 1 * x * 2
+@test :((a*b)^2) @simplify (a*b)^2
+#@test 
+ 
+
 @simplify (a^2*b^2)^3
-
-@simplify a + b + (0*c) + d
-@simplify a + b + (c*0) + d - d
-@simplify (a + d) - d
-
-@simplify a + b * c^0 + d
-@simplify a * x^y * b * x^d
-
-@simplify a * x^(12 + 3) * b * x^(42^3)
-
-@simplify (x+y)^(a*0) / (y+x)^0
+@test :() @simplify a + b + (0*c) + d
+@test :() @simplify a + b + (c*0) + d - d
+@test :() @simplify (a + d) - d
+@test :() @simplify a + b * c^0 + d
+@test :() @simplify a * x^y * b * x^d
+@test :() @simplify a * x^(12 + 3) * b * x^(42^3)
+@test :() @simplify (x+y)^(a*0) / (y+x)^0
 
 
 testt = @theory begin
     a + b       ==  b + a
     a + (b + c) ==  (a + b) + c
     d - d       => 0
-    a + -b      => a - b
-    a - b       => a + (-b)
+    a + neg(b)      => a - b
+    a - b       => a + neg(b)
 end
+
+ex = :((a + d) + neg(d))
+g = EGraph(ex)
+saturate!(g, testt)
 
[…]
