### 2021-03-19 a1c6a3a
fixed pattern matcher bug #40 and other minor improvements
