### 2021-03-19 ef24a63
architecture.md and some fixes to tests
