### 2021-03-19 32ffe44
architecture.md and some fixes to tests
