### 2021-03-19 #3078 @philzook58
Everything looks good. Two more comments:

1. I've been added as a reviewer twice in the paper draft
2.  I suspect the code is not quite displaying right in the draft for me. In particular, i think the \cup occurring in the theory combination is not rendering. Maybe I'm mistaken about the syntax. @0x0f0f0f should double check if there is anything else that looks off there

With the two points addressed, I think I consider my review complete
