### 2021-03-19 #3078 @0x0f0f0f
Yep. The cup is not rendering. Will fix with `union(...)` asap!
