### 2021-03-20 d9e7474
refactoring for readability. additional test suite for custom types with TermInterface
