### 2021-03-20 e23766b
refactoring for readability. additional test suite for custom types with TermInterface
