+++ src/EGraphs/saturation_report.jl
@@ -54,7 +54,7 @@ function Base.show(io::IO, x::Report)
     println(io, "\tStop Reason: $(x.reason)")
     println(io, "\tEGraph Saturated: $(x.saturated)")
     println(io, "\tIterations: $(x.iterations)")
-    println(io, "\tEGraph Size: $(length(g.M)) eclasses, $(length(g.H)) nodes")
+    println(io, "\tEGraph Size: $(length(g.emap)) eclasses, $(length(g.hashcons)) nodes")
     println(io, "\tTotal Time: $(x.total_time)")
     println(io, "\tSearch Time: $(x.search_stats)")
     println(io, "\tApply Time: $(x.apply_stats)")
