+++ test/test_mu.jl
@@ -4,7 +4,7 @@
 miu = @theory begin
     # Composition of the string monoid is associative
     x ⋅ (y ⋅ z) => (x ⋅ y) ⋅ z
-    # Add a U to the end of any string ending in I
+    # Add a uf to the end of any string ending in I
     x ⋅ :I =>  x ⋅ :I ⋅ :U
     # Double the string after the M
     :M ⋅ x  => :M ⋅ x ⋅ x
