### 2021-03-21 841312c
refactoring. major optimization improvements. version bump. removed autohashequals dep
