+++ test/test_mu.jl
@@ -5,18 +5,22 @@ miu = @theory begin
     # Composition of the string monoid is associative
     x ⋅ (y ⋅ z) => (x ⋅ y) ⋅ z
     # Add a uf to the end of any string ending in I
-    x ⋅ :I =>  x ⋅ :I ⋅ :U
+    x ⋅ :I ⋅ :END =>  x ⋅ :I ⋅ :U ⋅ :END
     # Double the string after the M
-    :M ⋅ x  => :M ⋅ x ⋅ x
+    :M ⋅ x ⋅ :END => :M ⋅ x ⋅ x ⋅ :END
     # Replace any III with a U
-    x ⋅ :I ⋅ :I ⋅ :I ⋅ y => x ⋅ :U ⋅ y
+    :I ⋅ :I ⋅ :I => :U 
     # Remove any UU
-    x ⋅ :U ⋅ :U ⋅ y => x ⋅ y
+    x ⋅ :U ⋅ :U ⋅ y => x ⋅ y 
 end
 
 
 @testset "MU puzzle" begin
     # no matter the timeout we set here,
-    # MU is not a theorem of the MIU system
-    @test false == areequal(miu, :(M⋅U), :(M⋅I), timeout=12, sizeout=2^13)
+    # MU is not a theorem of the MIU system 
+    params = SaturationParams(timeout=12, sizeout=2^13)
+    start = :(M⋅I⋅END)
+    g = EGraph(start)
+    saturate!(g, miu)
+    @test false == areequal(g, miu, start, :(M⋅U⋅END); params=params)
 end
