+++ test/runtests.jl
@@ -10,7 +10,7 @@ using Test
 # Metatheory.options[:verbose] = true
 # Metatheory.options[:printiter] = true
 
-@metatheory_init
+@metatheory_init ()
 
 Test.FallbackTestSet(desc) = Test.FallbackTestSet()
 
@@ -24,7 +24,7 @@ ts = Test.FallbackTestSet
       include("test_analysis.jl")
       include("test_extraction.jl")
       include("test_dynamic_ematch.jl")
-      # FIXME include("test_mu.jl")
+      include("test_mu.jl")
       include("test_boson.jl")
       include("test_while_interpreter.jl")
       include("test_theories.jl")
