### 2021-03-21 d96a83f
refactoring. major optimization improvements. version bump. removed autohashequals dep
