### 2021-03-21 cbf86f6
removed debug check
