### 2021-03-22 #3078 @dpsanders
👋 @philzook58, @mwillsey, @jpfairbanks: Many thanks for your comments. Could you please confirm here that you are happy with the paper and repository as is?
