### 2021-03-23 2189276
some fixes for extraction and differentiation example in CAS test
