+++ test/test_cas.jl
@@ -109,9 +109,9 @@ minus_t = @theory begin
     a - a       => 0
     a - b       => a + (-1*b)
     -a          => -1 * a
-    # TODO FIXME this rules bugs the pattern matcher
     a + (-b)    => a + (-1*b)
 end
+
 
 mulplus_t = @theory begin
     # TODO FIXME this rules improves performance and avoids commutative
@@ -131,13 +131,15 @@ pow_t = @theory begin
     0^x         => 0
     1^x         => 1
     x^1         => x
+    x * x       => x^2
     inv(x)      == x^(-1)
 end
 
 div_t = @theory begin
     x / 1 => x
     # x / x => 1 TODO SIGN ANALYSIS
-    x * (x / y) => 1 / y
+    x / (x / y) => y
+    x * (y / x) => y
     x * (y / z) == (x * y) / z
     x^(-1)      == 1 / x
 end
@@ -164,30 +166,46 @@ fold_t = @theory begin
     a::Number / b::Number   |> a/b
 end
 
-
+diff_t = @theory begin
+    ∂(y, x::Symbol) |> begin 
+        z = extract!(_egraph, simplcost; root=y.id)
+        @show z
+        zd = differentiate(z, x)
+        @show zd
+        zd
+    end
+end
 
 cas = fold_t ∪ mult_t ∪ plus_t ∪ minus_t ∪
-    mulplus_t ∪ pow_t ∪ div_t ∪ trig_t
+    mulplus_t ∪ pow_t ∪ div_t ∪ trig_t ∪ diff_t
+
+
 
 using Metatheory.TermInterface
 
+# ==================================================================
 ## Canonicalization
 
 
 function customlt(x,y)
-    if typeof(x) == Expr && Expr == typeof(y)
+    if typeof(x) == Expr && typeof(y) == Expr 
         false
     elseif typeof(x) == typeof(y)
         isless(x,y)
     elseif x isa Symbol && y isa Number
         false
-    else
-        true
-    end
+    elseif x isa Expr && y isa Number
+        false
+    elseif x isa Expr && y isa Symbol
+        false 
+    else true end
 end
 
 canonical_t = @theory begin
     # restore n-arity
+    (x * x)             => x^2
+    (x^n::Number * x)   => x^(n+1)
+    (x * x^n::Number)   => x^(n+1) 
     (x + (+)(ys...)) => +(x,ys...)
     ((+)(xs...) + y) => +(xs..., y)
     (x * (*)(ys...)) => *(x,ys...)
@@ -197,31 +215,58 @@ canonical_t = @theory begin
     (+)(xs...)      |> Expr(:call, :+, sort!(xs; lt=customlt)...)
 end
 
-Metatheory.options[:verbose] = false
-Metatheory.options[:printiter] = false
+Metatheory.options[:verbose] = true
+Metatheory.options[:printiter] = true
 
-function simplify(ex)
-    # rep = @timev begin
-        g = EGraph(ex)
-        params = SaturationParams(
-            scheduler=ScoredScheduler,
-            timeout=7,
-            schedulerparams=(8,2, Schedulers.exprsize),
-            #stopwhen=stopwhen,
-        )
-        saturate!(g, cas, params)
-        res = extract!(g, astsize)
-        # println(res)
-        # for (id, ec) ∈ g.emap
-        #     println(id, " => ", collect(ec.nodes))
-        #     println("\t\t", getdata(ec, ExtractionAnalysis{astsize}))
-        # end
-        res = rewrite(res, canonical_t; clean=false, m=@__MODULE__)
[…]
