### 2021-03-24 bda768c
dedicated types for global options and reasons of saturation halting. some minor improvements to tree walking
