+++ src/EGraphs/abstractanalysis.jl
@@ -4,10 +4,10 @@ abstract type AbstractAnalysis end
 
 # modify!(analysis::Type{<:AbstractAnalysis}, eclass::EClass) =
 #     error("Analysis does not implement modify!")
+islazy(an::Type{<:AbstractAnalysis})::Bool = false
 modify!(analysis::Type{<:AbstractAnalysis}, g, id) = nothing
 join(analysis::Type{<:AbstractAnalysis}, a, b) =
     error("Analysis does not implement join")
 make(analysis::Type{<:AbstractAnalysis}, g, a) =
     error("Analysis does not implement make")
 
-islazy(an::Type{<:AbstractAnalysis})::Bool = false
