### 2021-03-24 6a07aa9
dedicated types for global options and reasons of saturation halting. some minor improvements to tree walking
