+++ test/runtests.jl
@@ -7,8 +7,8 @@ using Metatheory.Util
 using DataStructures
 using Test
 
-# Metatheory.options[:verbose] = true
-# Metatheory.options[:printiter] = true
+Metatheory.options.verbose = false
+Metatheory.options.printiter = false
 
 @metatheory_init ()
 
