+++ src/options.jl
@@ -1,0 +1,18 @@
+using Parameters
+
+"""
+Global configurable options for Metatheory.
+"""
+@with_kw mutable struct Options 
+    """
+    Print or not information such as saturation reports 
+    """
+    verbose::Bool = false
+    """
+    Print iteration numbers in equality saturation
+    """
+    printiter::Bool = false
+end
+
+
+options = Options()
