### 2021-03-24 5bd289b
working on the new pattern interface
