+++ src/TermInterface.jl
@@ -10,6 +10,7 @@ gethead(e::Expr) = isexpr(e, :call) ? e.args[1] : e.head
 getargs(e::Expr) = @view e.args[(isexpr(e, :call) ? 2 : 1):end]
 istree(e::Expr) = true
 getmetadata(e::Expr) = (iscall=isexpr(e, :call),)
+arity(e::Expr) = length(getargs(e)) # optional
 preprocess(e::Expr) = cleanast(e)
 
 # Fallback implementation for all other types
@@ -17,6 +18,7 @@ gethead(e) = e
 getargs(e) = []
 istree(a) = false
 getmetadata(e) = nothing
+arity(e) = 0 # optional
 preprocess(e) = e
 
 export gethead
@@ -24,4 +26,5 @@ export getargs
 export istree
 export getmetadata
 export preprocess
+export arity
 end
