### 2021-03-25 831209d
completed new pattern interface. eclass unioning improvements
