+++ src/Rules/Rules.jl
@@ -1,6 +1,7 @@
 module Rules
     using ..TermInterface
     using ..Util
+    import ..closure_generator
 
     include("patterns.jl")
     include("rule_types.jl")
