### 2021-03-25 0838426
completed new pattern interface. eclass unioning improvements
