+++ src/EGraphs/saturation.jl
@@ -1,4 +1,4 @@
-const Match = Tuple{Rule, Sub, Int64} #bool isright
+const Match = Tuple{Rule, Union{Nothing, Pattern}, Sub, Int64}
 const MatchesBuf = OrderedSet{Match}
 # const MatchesBuf = Dict{Rule,Set{Sub}}
 
@@ -8,52 +8,34 @@ import ..@log
 
 using .Schedulers
 
-"""
-Exactly like [`addexpr!`](@ref), but instantiate pattern variables
-from a substitution, resulting from a pattern matcher run.
-"""
-function addexprinst!(G::EGraph, spat, sub::Sub, side::Symbol, r::Rule)::EClass
-    # e = preprocess(pat)
-    # println("========== $pat ===========")
 
-    spat isa EClass && return spat
-
-    pat = preprocess(spat)
-
+function instantiate(pat::PatVar, sub::Sub, rule::Rule)
     if haskey(sub, pat)
         (eclass, lit) = sub[pat]
-        pat = (lit !== nothing ? lit : eclass)
-    elseif pat isa Symbol
-        # TODO tell me the rule!
-        # for (k,v) ∈ sub
-        #     println("$k => $v"); println()
-        # end
-        # println.(sort(collect(G.emap); lt=((x,y)-> x[1] < y[1])))
-        error("unbound pattern variable $pat in rule $r")
-    elseif side == :right && pat isa QuoteNode && pat.value isa Symbol
-        pat = pat.value
+        return (lit !== nothing ? lit : eclass)
+    else
+        error("unbound pattern variable $pat in rule $rule")
     end
-
-    # println("pat $pat")
-    pat isa EClass && return pat
-
-    if istree(pat)
-        args = getargs(pat)
-        n = length(args)
-        class_ids = Vector{Int64}(undef, length(args))
-        for i ∈ 1:n
-            # println("child $child")
-            @inbounds child = args[i]
-            c_eclass = addexprinst!(G, child, sub, side, r)
-            @inbounds class_ids[i] = c_eclass.id
-        end
-        node = ENode(pat, class_ids)
-        return add!(G, node)
-    end
-
-    return add!(G, ENode(pat))
 end
 
+function instantiate(pat::PatLiteral, sub::Sub, rule::Rule)
+    pat.val
+end
+
+function instantiate(pat::PatTypeAssertion, sub::Sub, rule::Rule)
+    instantiate(pat.var, sub, rule)
+end
+
+function instantiate(pat::PatTerm, sub::Sub, rule::Rule)
+    # TODO support custom types here!
+    # similarterm ? ask Shashi
+    meta = pat.metadata
+    if meta !== nothing && meta.iscall
+        Expr(:call, pat.head, map(x -> instantiate(x, sub, rule), pat.args)...)
+    else
+        Expr(pat.head, map(x -> instantiate(x, sub, rule), pat.args)...)
+    end
+end
 
 
 function cached_ids(egraph::EGraph, side)#::Vector{Int64}
@@ -69,7 +51,28 @@ function cached_ids(egraph::EGraph, side)#::Vector{Int64}
     end
 end
 
-function eqsat_search!(egraph::EGraph, theory::Vector{Rule},
+function search_rule!(g::EGraph, r::SymbolicRule, id::Int64, matches::MatchesBuf)
+    for sub in ematch(g, r.left, id)
+        !isempty(sub) && push!(matches, (r, r.right, sub, id))
+    end
+end
+
+function search_rule!(g::EGraph, r::DynamicRule, id::Int64, matches::MatchesBuf)
+    for sub in ematch(g, r.left, id)
[…]
