### 2021-03-25 5f78f28
function barriers for rule application
