+++ src/EGraphs/saturation.jl
@@ -101,83 +101,95 @@ function eqsat_search!(egraph::EGraph, theory::Vector{<:Rule},
     return matches
 end
 
+function apply_rule!(g::EGraph, rule::UnequalRule,
+        match::Match, matches::MatchesBuf, rep::Report, mod::Module)
+    (_, pat, sub, id) = match
+    lc = id
+    rinst = instantiate(pat, sub, rule)
+    rc = addexpr!(g, rinst).id
+    delete!(matches, match)
+    if find(g, lc) == find(g, rc)
+        @log "Contradiction!" rule
+        rep.reason = Contradiction()
+        return (false, rep)
+    end
+    return (true, nothing)
+end
+
+function apply_rule!(g::EGraph, rule::SymbolicRule, 
+        match::Match, matches::MatchesBuf, rep::Report,  mod::Module)
+    (_, pat, sub, id) = match
+    rinst = instantiate(pat, sub, rule)
+    rc = addexpr!(g, rinst).id
+    merge!(g, id, rc)
+    return (true, nothing)
+end
+
+function apply_rule!(g::EGraph, rule::DynamicRule, 
+        match::Match, matches::MatchesBuf, rep::Report,  mod::Module)
+    (_, pat, sub, id) = match
+    lc = id
+    f = Rules.getrhsfun(rule, mod)
+    actual_params = map(rule.patvars) do x
+        (eclass, literal) = sub[x]
+        literal !== nothing ? literal : eclass
+    end
+    r = f(geteclass(g, lc), g, actual_params...)
+    rc = addexpr!(g, r)
+    merge!(g,lc,rc.id)
+    return (true, nothing)
+end
+
+
 using .ReportReasons
-function eqsat_apply!(egraph::EGraph, matches::MatchesBuf,
-        scheduler::AbstractScheduler, report::Report,
+function eqsat_apply!(g::EGraph, matches::MatchesBuf,
+        scheduler::AbstractScheduler, rep::Report,
         mod::Module, params::SaturationParams)::Report
     i = 0
     for match ∈ matches
-        (rule, pat, sub, id) = match
         i += 1
 
         if i % 300 == 0
-            if params.sizeout > 0 && length(egraph.uf) > params.sizeout
+            if params.sizeout > 0 && length(g.uf) > params.sizeout
                 @log "E-GRAPH SIZEOUT"
-                report.reason = EClassLimit()
-                return report
+                rep.reason = EClassLimit()
+                return rep
             end
 
             if params.stopwhen()
                 @log "Halting requirement satisfied"
-                report.reason = ConditionSatisfied()
-                return report
+                rep.reason = ConditionSatisfied()
+                return rep
             end
         end
 
+        rule=match[1]
         writestep!(scheduler, rule)
-
-        if rule isa UnequalRule
-            lc = id
-            rinst = instantiate(pat, sub, rule)
-            rc = addexpr!(egraph, rinst).id
-            delete!(matches, match)
-            if find(egraph, lc) == find(egraph, rc)
-                    @log "Contradiction!" rule
-                    report.reason = Contradiction()
-                    return report
-            end
-        elseif rule isa SymbolicRule 
-            lc = id
-            rinst = instantiate(pat, sub, rule)
-            rc = addexpr!(egraph, rinst).id
-
-                merge!(egraph, lc, rc)
-            # end
-        elseif rule isa DynamicRule # execute the right hand!
-            lc = id
-            f = Rules.getrhsfun(rule, mod)
-            actual_params = map(rule.patvars) do x
-                (eclass, literal) = sub[x]
[…]
