### 2021-03-25 5fffc1d
function barriers for rule application
