### 2021-03-25 0945a9c
fixed eclass limit. renaming egraph attributes for convenience
