+++ src/EGraphs/saturation_params.jl
@@ -6,7 +6,7 @@ Configurable Parameters for the equality saturation process.
 @with_kw mutable struct SaturationParams
     timeout::Int = 7
     # default sizeout. TODO make this bytes
-    sizeout::Int = 2^14
+    # sizeout::Int = 2^14
     matchlimit::Int = 5000
     eclasslimit::Int = 5000
     stopwhen::Function = ()->false
