### 2021-03-25 434093f
fixed eclass limit. renaming egraph attributes for convenience
