+++ src/EGraphs/saturation_report.jl
@@ -56,7 +56,7 @@ function Base.show(io::IO, x::Report)
     println(io, "=================")
     println(io, "\tStop Reason: $(x.reason)")
     println(io, "\tIterations: $(x.iterations)")
-    println(io, "\tEGraph Size: $(length(g.emap)) eclasses, $(length(g.hashcons)) nodes")
+    println(io, "\tEGraph Size: $(g.numclasses) eclasses, $(length(g.memo)) nodes")
     println(io, "\tTotal Time: $(x.total_time)")
     println(io, "\tSearch Time: $(x.search_stats)")
     println(io, "\tApply Time: $(x.apply_stats)")
