+++ test/test_egraphs.jl
@@ -22,20 +22,20 @@ end
     t1 = addexpr!(G, :b)
 
     t2 = addexpr!(G, :c)
-    # display(G.emap); println()
+    # display(G.classes); println()
 
     c_id = merge!(G, t2.id, t1.id)
-    # display(G.emap); println()
+    # display(G.classes); println()
     @test in_same_set(G.uf, c_id, t1.id)
     @test in_same_set(G.uf, t2.id, t1.id)
     # println(find_root!(G.uf, t2.id))
     @test find_root!(G.uf, t2.id) == 4
     rebuild!(G)
     # f(a,b) = f(a,c)
-    # display(G.emap); println()
-    # display(G.hashcons); println()
+    # display(G.classes); println()
+    # display(G.memo); println()
 
-    # for (id, ec) ∈ G.emap
+    # for (id, ec) ∈ G.classes
     #     println(id)
     #     dump.(ec.nodes)
     # end
@@ -56,11 +56,11 @@ end
     c_id = merge!(G, t1.id, 1) # a == apply(6,f,a)
     c2_id = merge!(G, t2.id, 1) # a == apply(9,f,a)
 
-    # display(G.emap); println()
+    # display(G.classes); println()
 
     rebuild!(G)
 
-    # display(G.emap); println()
+    # display(G.classes); println()
 
     t3 = addexpr!(G, apply(3, f, :a))
     t4 = addexpr!(G, apply(7, f, :a))
