### 2021-03-26 f823eb8
multipattern madness!
