+++ src/EGraphs/saturation.jl
@@ -38,39 +38,85 @@ function instantiate(pat::PatTerm, sub::Sub, rule::Rule)
 end
 
 
-function cached_ids(egraph::EGraph, side)#::Vector{Int64}
-    # outermost symbol in rule side
-    # side = remove_assertions(side)
-    # side = unquote_sym(side)
-    if istree(side)
-        sym = gethead(side)
-        get(egraph.symcache, sym, [])
-    else
-        # collect(keys(egraph.classes))
-        keys(egraph.classes)
-    end
+function cached_ids(g::EGraph, p::Pattern)#::Vector{Int64}
+    keys(g.classes)
 end
+
+# FIXME 
+# function cached_ids(g::EGraph, p::PatTerm)
+#     get(g.symcache, p.head, [])
+# end
+
+# function cached_ids(g::EGraph, p::PatLiteral)
+#     get(g.symcache, p.val, [])
+# end
 
 function search_rule!(g::EGraph, r::SymbolicRule, id::Int64, matches::MatchesBuf)
     for sub in ematch(g, r.left, id)
-        !isempty(sub) && push!(matches, (r, r.right, sub, id))
+        isempty(sub) && error("empty sub")
+        push!(matches, (r, r.right, sub, id))
     end
 end
 
 function search_rule!(g::EGraph, r::DynamicRule, id::Int64, matches::MatchesBuf)
     for sub in ematch(g, r.left, id)
-        !isempty(sub) && push!(matches, (r, nothing, sub, id))
+        isempty(sub) && error("empty sub")
+        push!(matches, (r, nothing, sub, id))
     end
 end
 
 function search_rule!(g::EGraph, r::BidirRule, id::Int64, matches::MatchesBuf)
     for sub in ematch(g, r.left, id)
-        !isempty(sub) && push!(matches, (r, r.right, sub, id))
+        isempty(sub) && error("empty sub")
+        push!(matches, (r, r.right, sub, id))
     end
     for sub in ematch(g, r.right, id)
-        !isempty(sub) && push!(matches, (r, r.left, sub, id))
+        isempty(sub) && error("empty sub")
+        push!(matches, (r, r.left, sub, id))
     end
 end
+
+function Base.show(io::IO, s::Sub)
+    print(io, "Sub[")
+    kvs = collect(s)
+    n = length(kvs)
+    for i ∈ 1:n
+        print(io, kvs[i][1], " => ", kvs[i][2][1].id)
+        if i < n 
+            print(io, ",")
+        end
+    end
+    print(io, "]")
+end
+
+function search_rule!(g::EGraph, r::MultiPatRewriteRule, id::Int64, matches::MatchesBuf)
+    buf = ematch(g, r.left, id)
+    if isempty(buf)
+        return 
+    end
+    pats_todo = reverse(copy(r.pats))
+    while !isempty(pats_todo)
+        pat = pop!(pats_todo)
+        # println("====================")
+        # @show pat
+        ids = cached_ids(g, pat)
+        newbuf = SubBuf()
+        while !isempty(buf)
+            sub = pop!(buf)
+            # @show sub
+            isempty(sub) && continue
+            for i ∈ ids
+                ematch(g, pat, i; sub=sub, buf=newbuf)
+            end
+        end
+        buf = copy(newbuf)
+    end
+    for sub in buf
+        # println("FINALLY ", sub, " $id")
+        push!(matches, (r, r.right, sub, id))
+    end
+end
+
 
 function eqsat_search!(egraph::EGraph, theory::Vector{<:Rule},
         scheduler::AbstractScheduler)::MatchesBuf
