### 2021-03-26 6058afe
multipattern madness!
