+++ src/Rules/exports.jl
@@ -5,7 +5,9 @@ export PatVar
 export PatTerm
 export PatTypeAssertion
 export PatSplatVar
+export PatEquiv
 export patvars
+export @pat
 
 # include("rules/rule_types.jl")
 export Rule
@@ -15,6 +17,7 @@ export BidirRule
 export EqualityRule
 export UnequalRule
 export DynamicRule
+export MultiPatRewriteRule
 
 
 # include("rules/rule_dsl.jl")
