+++ src/EGraphs/EGraphs.jl
@@ -30,9 +30,11 @@ export geteclass
 export arity
 export EGraph
 export merge!
+export in_same_class
 export addexpr!
 export addanalysis!
 export rebuild!
+
 
 include("analysis.jl")
 export analyze!
