### 2021-03-27 1377b63
added PatAllTerm, restructured tests
