### 2021-03-27 ca41f02
added PatAllTerm, restructured tests
