+++ src/Rules/Rules.jl
@@ -4,6 +4,7 @@ module Rules
     import ..closure_generator
 
     include("patterns.jl")
+    include("patterns_syntax.jl")
     include("rule_types.jl")
     include("rule_cache.jl")
     include("rule_dsl.jl")

[…]
