+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"api/#API-Documentation","page":"API Documentation","title":"API Documentation","text":"","category":"section"},{"location":"api/#Metatheory","page":"API Documentation","title":"Metatheory","text":"","category":"section"},{"location":"api/","page":"API Documentation","title":"API Documentation","text":"Modules = [Metatheory]","category":"page"},{"location":"api/#Metatheory.Options","page":"API Documentation","title":"Metatheory.Options","text":"mutable struct Options\n\nGlobal configurable options for Metatheory.\n\n\n\nFields\n\nverbose::Bool\nPrint or not information such as saturation reports  Default: false\nprintiter::Bool\nPrint iteration numbers in equality saturation  Default: false\n\n\n\n\n\n","category":"type"},{"location":"api/#Metatheory.Classic","page":"API Documentation","title":"Metatheory.Classic","text":"","category":"section"},{"location":"api/","page":"API Documentation","title":"API Documentation","text":"CurrentModule = Metatheory\nDocTestSetup  = quote\n    using Metatheory\n    using Metatheory.EGraphs\nend","category":"page"},{"location":"api/","page":"API Documentation","title":"API Documentation","text":"Modules = [Classic]","category":"page"},{"location":"api/#Metatheory.Classic","page":"API Documentation","title":"Metatheory.Classic","text":"This module contains classical rewriting functions and utilities\n\n\n\nImports\n\nBase\nBase.Meta\nCore\nMatchCore\nMetatheory.Rules\nMetatheory.Util\n\n\n\n\n\n","category":"module"},{"location":"api/#Metatheory.Classic.compile_theory-Tuple{Array{var\"#s13\",1} where var\"#s13\"<:Rule,Module}","page":"API Documentation","title":"Metatheory.Classic.compile_theory","text":"Compile a theory to a closure that does the pattern matching job Returns a RuntimeGeneratedFunction, which does not use eval and is as fast as a regular Julia anonymous function 🔥\n\n\n\n\n\n","category":"method"},{"location":"api/#Metatheory.Classic.rewrite-Tuple{Any,Union{Function, Array{var\"#s20\",1} where var\"#s20\"<:Rule}}","page":"API Documentation","title":"Metatheory.Classic.rewrite","text":"This function executes a classical rewriting algorithm on a Julia expression ex. Classical rewriting applies rule in order with a fixed point iteration:\n\nThis algorithm heavily relies on RuntimeGeneratedFunctions.jl and the MatchCore pattern matcher. NOTE: this does not involve the use of EGraphs.EGraph or equality saturation (EGraphs.saturate!). When using rewrite, be aware of infinite loops: Since rules are matched in order in every iteration, it is possible that commonly used symbolic rules such as commutativity or associativity of operators may cause this algorithm to have a cycling computation instantly. This algorithm detects cycling computation by keeping an history of hashes, and instantly returns when a cycle is detected.\n\nThis algorithm is suitable for simple, deterministic symbolic rewrites. For more advanced u
[…]
