### 2021-03-28 bed4269
cleaned up repo, architecture details, readme improvements
