### 2021-03-28 c21687b
cleaned up repo, architecture details, readme improvements
