### 2021-03-28 9f89f85
cleaner rule constructor, imperceivably faster ematch
