### 2021-03-28 ad881d7
cleaner rule constructor, imperceivably faster ematch
