+++ src/EGraphs/ematch.jl
@@ -22,21 +22,21 @@ the remaining E-nodes. The base case of this recursion is the empty list, which
 requires no extension to the substitution; the other case relies on Match to find the
 substitutions that match the first term to the first E-node.
 """
-function ematchlist(e::EGraph, t::AbstractVector{Pattern}, v::AbstractVector{Int64}, sub::Sub; buf=SubBuf())::SubBuf
+function ematchlist(e::EGraph, t::AbstractVector{Pattern}, v::AbstractVector{Int64}, sub::Sub, buf::SubBuf)::SubBuf
     if length(t) == 0 || length(v) == 0
         if !isempty(sub)
             push!(buf, sub)
         end
     else
-        for sub1 in ematchstep(e, t[1], v[1], sub)
-            ematchlist(e, (@view t[2:end]), (@view v[2:end]), sub1; buf=buf)
+        for sub1 in ematchstep(e, t[1], v[1], sub, SubBuf(), nothing)
+            ematchlist(e, (@view t[2:end]), (@view v[2:end]), sub1, buf)
         end
     end
     return buf
 end
 
 # Tries to match on a pattern variable
-function ematchstep(g::EGraph, t::PatVar, v::Int64, sub::Sub; lit=nothing, buf=SubBuf())::SubBuf
+function ematchstep(g::EGraph, t::PatVar, v::Int64, sub::Sub, buf::SubBuf, lit=nothing)::SubBuf
     if haskey(sub, t)
         if find(g, first(sub[t])) == find(g, v)
             push!(buf, sub)
@@ -48,7 +48,7 @@ function ematchstep(g::EGraph, t::PatVar, v::Int64, sub::Sub; lit=nothing, buf=S
 end
 
 # Tries to match on literals
-function ematchstep(g::EGraph, t::PatLiteral, v::Int64, sub::Sub; lit=nothing, buf=SubBuf())::SubBuf
+function ematchstep(g::EGraph, t::PatLiteral, v::Int64, sub::Sub, buf::SubBuf, lit=nothing)::SubBuf
     ec = geteclass(g, v)
     for n in ec
         if arity(n) == 0 && t.val == n.head
@@ -66,25 +66,26 @@ end
 
 
 # tries to match on type assertions
-function ematchstep(g::EGraph, t::PatTypeAssertion, v::Int64, sub::Sub; lit=nothing, buf=SubBuf())::SubBuf
+function ematchstep(g::EGraph, t::PatTypeAssertion, v::Int64, sub::Sub, buf::SubBuf, lit=nothing)::SubBuf
     ec = geteclass(g, v)
     for n in ec
         if arity(n) == 0
             !(typeof(n.head) <: t.type) && continue
-            ematchstep(g, t.var, v, sub; lit=n.head, buf=buf)
+            ematchstep(g, t.var, v, sub, buf, n.head)
             continue
         end
     end
     return buf
 end
 
-function ematchstep(g::EGraph, t::PatEquiv, v::Int64, sub::Sub; lit=nothing, buf=SubBuf())::SubBuf
+# PATEQUIV mechanism
+function ematchstep(g::EGraph, t::PatEquiv, v::Int64, sub::Sub, buf::SubBuf, lit=nothing)::SubBuf
     buf1 = SubBuf()
     buf2 = SubBuf()
 
     
-    for sub1 ∈ ematchstep(g, t.left, v, sub; buf=buf1)
-        ematchstep(g, t.right, v, sub1; buf=buf2)
+    for sub1 ∈ ematchstep(g, t.left, v, sub, buf1)
+        ematchstep(g, t.right, v, sub1, buf2)
     end
 
     if !isempty(buf1) && !isempty(buf2) 
[…]
