### 2021-03-29 b0889cf
better immutable unioning of eclasses
