### 2021-03-29 0346195
egrpah pruning rules!
