+++ test/cas/cas_theory.jl
@@ -12,7 +12,8 @@ mult_t = commutative_monoid(:(*), 1)
 plus_t = commutative_monoid(:(+), 0)
 
 minus_t = @theory begin
-    a - a       => 0
+    # TODO Jacques Carette's post in zulip chat
+    @pruningrule a - a       => 0
     a - b       => a + (-1*b)
     -a          => -1 * a
     a + (-b)    => a + (-1*b)
@@ -22,8 +23,8 @@ end
 mulplus_t = @theory begin
     # TODO FIXME this rules improves performance and avoids commutative
     # explosion of the egraph
-    0 * a       => 0
-    a * 0       => 0
+    @pruningrule 0 * a => 0
+    @pruningrule a * 0       => 0
     a * (b + c) == ((a*b) + (a*c))
     a + (b * a) => ((b+1)*a)
 end
@@ -33,9 +34,9 @@ pow_t = @theory begin
     x^n * x^m   == x^(n+m)
     (x * y)^z   == x^z * y^z
     (x^p)^q     == x^(p*q)
-    x^0         => 1
-    0^x         => 0
-    1^x         => 1
+    @pruningrule x^0         => 1
+    @pruningrule 0^x         => 0
+    @pruningrule 1^x         => 1
     x^1         => x
     x * x       => x^2
     inv(x)      == x^(-1)
@@ -65,11 +66,11 @@ end
 
 # Dynamic rules
 fold_t = @theory begin
-    -(a::Number)            |> -a
-    a::Number + b::Number   |> a + b
-    a::Number * b::Number   |> a * b
-    a::Number ^ b::Number   |> begin b < 0 && a isa Int && (a = float(a)) ; a^b end
-    a::Number / b::Number   |> a/b
+    @pruningrule -(a::Number)            |> -a
+    @pruningrule a::Number + b::Number   |> a + b
+    @pruningrule a::Number * b::Number   |> a * b
+    @pruningrule a::Number ^ b::Number   |> begin b < 0 && a isa Int && (a = float(a)) ; a^b end
+    @pruningrule a::Number / b::Number   |> a/b
 end
 
 using Calculus: differentiate
