+++ src/Rules/exports.jl
@@ -19,10 +19,11 @@ export EqualityRule
 export UnequalRule
 export DynamicRule
 export MultiPatRewriteRule
-
+export canprune
 
 # include("rules/rule_dsl.jl")
 export Rule
 export Theory
 export @rule
 export @theory
+export @pruningrule
