+++ test/cas/cas_simplify.jl
@@ -1,4 +1,3 @@
-## Theory for CAS
 using Metatheory
 using Metatheory.Library
 using Metatheory.EGraphs
