### 2021-03-29 97feba0
egrpah pruning rules!
