+++ docs/src/theories.md
@@ -1,4 +1,4 @@
-# Rules and Theories Syntax
+# Rules and Theories
 
 # Rule Syntax for Classical Rewriting
 
@@ -27,7 +27,12 @@ You are not limited to composing theories, you can
 manipulate and create them at both runtime and compile time
 as regular vectors.
 
-```julia
+```@example 2
+using Metatheory
+using Metatheory.EGraphs
+using Metatheory.Library
+
+comm_monoid = commutative_monoid(:(*), 1);
 comm_group = @theory begin
     a + 0 => a
     a + b => b + a
@@ -52,7 +57,7 @@ Dynamic rules, are similar to anonymous functions. Instead of a symbolic
 substitution, the right hand of a dynamic `|>` rule is evaluated during
 rewriting: the values that produced a match are bound to the pattern variables.
 
-```julia
+```@example 2
 fold_mul = @theory begin
     a::Number * b::Number |> a*b
 end
@@ -64,17 +69,25 @@ t = comm_monoid ∪ fold_mul
 ## Escaping
 
 You can escape values in the left hand side of rules using `$` just
-as you would do with the regular [quoting/unquoting]() mechanism.
+as you would do with the regular [quoting/unquoting](https://docs.julialang.org/en/v1/manual/metaprogramming/) mechanism.
 
 
-```julia
+```@example 2
 example = @theory begin
     a + $(3+2) |> :something
 end
 ```
 
-Becomes
+## Patterns
+
+```@autodocs
+Modules = [Rules]
+Pages = ["patterns.jl", "patterns_syntax.jl"]
 ```
-1-element Vector{Rule}:
- Rule(:(a + 5 |> :something))
+
+## Rules 
+
+```@autodocs
+Modules = [Rules]
+Pages = ["rule_types.jl", "rule_dsl.jl"]
 ```
