+++ docs/src/options.md
@@ -1,0 +1,10 @@
+# Global Metatheory.jl Options
+
+
+```@meta
+CurrentModule = Metatheory
+```
+
+```@autodocs
+Modules = [Metatheory]
+```
