+++ docs/src/schedulers.md
@@ -1,0 +1,6 @@
+# Scheduling for E-Graph rewriting
+
+
+```@autodocs
+Modules = [Metatheory.EGraphs.Schedulers]
+```
