+++ docs/src/api.md
@@ -1,65 +1,0 @@
-# API Documentation
-
-## Metatheory
-
-```@autodocs
-Modules = [Metatheory]
-```
-
-## Metatheory.Classic
-
-```@meta
-CurrentModule = Metatheory
-DocTestSetup  = quote
-    using Metatheory
-    using Metatheory.EGraphs
-end
-```
-
-
-```@autodocs
-Modules = [Classic]
-```
-
-
-## Metatheory.EGraphs
-
-```@meta
-CurrentModule = Metatheory
-DocTestSetup  = quote
-    using Metatheory
-end
-```
-
-
-```@autodocs
-Modules = [Metatheory.EGraphs]
-```
-
-## Metatheory.EGraphs.Schedulers
-
-```@meta
-CurrentModule = Metatheory
-DocTestSetup  = quote
-    using Metatheory
-end
-```
-
-
-```@autodocs
-Modules = [Metatheory.EGraphs.Schedulers]
-```
-
-## Metatheory.Util
-
-```@meta
-CurrentModule = Metatheory
-DocTestSetup  = quote
-    using Metatheory
-end
-```
-
-
-```@autodocs
-Modules = [Util]
-```
