### 2021-03-29 #3078 @dpsanders
@0x0f0f0f: Please add DOIs for all the references.
