### 2021-03-29 #3078 @dpsanders
@0x0f0f0f: Please add an explanation for the meaning of Figure 1. What do the arrows and dashed boxes represent? I suggest you also lightly shade or colour what is new in each subsequent part of the figure.
