### 2021-03-29 #3078 @0x0f0f0f
@dpsanders some references are github repos and do not have DOIs. Do I have to remove them from `.bib` file and move the links to the paper body?
