### 2021-03-29 #3078 @0x0f0f0f
@dpsanders added explanation and coloring to the figure
