### 2021-03-29 7667ecd
Minor grammar and style fixes
