+++ paper/paper.bib
@@ -172,6 +172,7 @@ series = {PLDI '15}
   title={Equality Saturation for Tensor Graph Superoptimization},
   author={Yang, Yichen and Phothilimtha, Phitchaya Mangpo and Wang, Yisu Remy and Willsey, Max and Roy, Sudip and Pienaar, Jacques},
   journal={arXiv preprint arXiv:2101.01332},
+  url={https://arxiv.org/pdf/2101.01332.pdf},
   year={2021}
 }
 
