+++ paper/paper.bib
@@ -196,14 +196,14 @@ numpages = {9}
 }
 
 @article{philzuck1,
-  title={E-graphs in Julia (Part I)},
+  title={E-graphs in Julia {Part I}},
   author={Zucker, Philip},
   year={2020},
   url = {https://www.philipzucker.com/egraph-1/}
 }
 
 @article{philzuck2,
-  title={E-Graph Pattern Matching (Part II)},
+  title={E-Graph Pattern Matching {Part II}},
   author={Zucker, Philip},
   year={2020},
   url={https://www.philipzucker.com/egraph-2/}
