+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"theories/#Rules-and-Theories","page":"Rules and Theories","title":"Rules and Theories","text":"","category":"section"},{"location":"theories/#Rule-Syntax-for-Classical-Rewriting","page":"Rules and Theories","title":"Rule Syntax for Classical Rewriting","text":"","category":"section"},{"location":"theories/","page":"Rules and Theories","title":"Rules and Theories","text":"Kind Supported in Left Hand Side Operator Supported in Right Hand Side\nSymbolic  Rule x  (pattern variables)  :foo (symbol literals)  x::Type (type assertions)  $(2 + 3) (unquoting)  a... (pattern variable destructuring, matches many subterms as a tuple)   Other literals are supported. => x (pattern variables)  :foo(symbol literals)  a... (pattern variable destructuring)   $(2 + 3) (unquoting)  Other literals are supported.\nDynamic Rule Same as above |> Dynamic rules can execute all valid Julia code. The pattern variables  that matched are available (bound) in the r.h.s.. Other global variables  in the execution module are bound. An additional variable _lhs_expr is bound, referring to the left hand side that matched the rule.\nEquational Rule Unsupported == Unsupported","category":"page"},{"location":"theories/#Rule-Syntax-for-EGraphs-Rewriting","page":"Rules and Theories","title":"Rule Syntax for EGraphs Rewriting","text":"","category":"section"},{"location":"theories/","page":"Rules and Theories","title":"Rules and Theories","text":"Kind Supported in  Left Hand Side Operator Supported in Right Hand Side\nSymbolic  Rule x  (pattern variables)  :foo (symbol literals)  x::Type (type assertions)  $(2 + 3) (unquoting)   Other literals are supported. Pattern variable destructuring is not supported. => x (pattern variables)  :foo(symbol literals)   $(2 + 3) (unquoting)  Other literals are supported.\nDynamic Rule Same as above |> Dynamic rules execute valid Julia code. The pattern variables  that matched are available (bound) in the r.h.s.. Other global variables  in the execution module are bound. An additional variable _lhs_expr is bound,  referring to the left hand side that matched the rule.  NOTE: additionally, the _egraph variable is bound,  referring to the current EGraph on which rewriting is happening.\nEquational Rule Same as Symbolic Rules. == Same as left hand side of symbolic rules.","category":"page"},{"location":"theories/#Theories-are-Collections-and-Composable","page":"Rules and Theories","title":"Theories are Collections and Composable","text":"","category":"section"},{"location":"theories/","page":"Rules and Theories","title":"Rules and Theories","text":"Theories are just collections, precisely vectors of the Rule object, and can be composed as regular Julia collections. The most useful way of composing theories is unioning them with the '∪' operator. You are not limited to composing theories, you can manipulate and create them at both runtime and compile time as regular vectors.","category":"page"},{"location":"theories/","page":"Rules and Theories","title":"Rules and Theories","text":"using Metatheory\nusing Metatheory.EGraphs\nusing Metatheory.Library\n\ncomm_monoid = commutative_monoid(:(*), 1);\ncomm_group = @theory begin\n    a + 0 => a\n    a + b => b + a\n    a + inv(a) => 0 # inverse\n    a + (b + c) => (a + b) + c\nend\ndistrib = @theory begin\n    a * (b + c) => (a * b) + (a * c)\nend\nt = comm_monoid ∪ comm_group ∪ distrib","category":"page"},{"location":"theories/#Type-Assertions-and-Dynamic-Rules","page":"Rules and Theories","title":"Type Assertions and Dynamic Rules","text":"","category":"s
[…]
