### 2021-03-30 #3078 @kthyng
@0x0f0f0f Starting to read your paper. I don't see a reason to put your position in your author list ("undergraduate student"). Do you have a special reason to include it? If not I suggest removing it. Also is there is a reason you have an additional label as "independent researcher"? If you are attending University of Pisa, you could just list that, though you should expand the information to include at least the country.
