### 2021-03-30 #3078 @kthyng
@0x0f0f0f I wasn't able to finish reading your paper earlier, sorry. But I've looked through it now and just a few more comments:
1. Please differentiate between references that are cited inline, in which case you should remove the [] from the reference, and references that are parenthetical (usually at the end of the sentence) which should keep the [] around the reference. You can get more info [here](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography).
2. At least a couple references in the reference list look like they should have "Julia" capitalized (i think it is supposed to be capitalized?). You can preserve capitalization from the .bib file by placing {} around those words.
