### 2021-03-30 edce5d9
redesigned ematchlist, custom disjoint sets data structure, multithreaded search, bugfixes
