### 2021-03-30 5b7f7fd
redesigned ematchlist, custom disjoint sets data structure, multithreaded search, bugfixes
