+++ src/EGraphs/saturation_search.jl
@@ -1,15 +1,34 @@
-const Match = Tuple{Rule, Union{Nothing, Pattern}, Sub, Int64}
+const Match = Tuple{Rule,Union{Nothing,Pattern},Sub,Int64}
 const MatchesBuf = Vector{Match}
 
-function cached_ids(g::EGraph, p::Pattern)#::Vector{Int64}
+function cached_ids(g::EGraph, p::Pattern)# ::Vector{Int64}
     keys(g.classes)
 end
 
 # FIXME 
 function cached_ids(g::EGraph, p::PatTerm)
-    # println(p)
-    # println(get(g.symcache, p.head, []))
+    # println("pattern $p, $(p.head)")
+    # println("all ids")
+    # keys(g.classes) |> println
+    # println("cached symbols")
+    # cached = get(g.symcache, p.head, Set{Int64}())
+    # println("symbols where $(p.head) appears")
+    # appears = Set{Int64}() 
+    # for (id, class) ∈ g.classes 
+    #     for n ∈ class 
+    #         if n.head == p.head
+    #             push!(appears, id) 
+    #         end
+    #     end
+    # end
+    # # println(appears)
+    # if !(cached == appears)
+    #     @show cached 
+    #     @show appears
+    # end
+
     keys(g.classes)
+    # cached
     # get(g.symcache, p.head, [])
 end
 
@@ -17,24 +36,35 @@ end
 #     get(g.symcache, p.val, [])
 # end
 
-function search_rule!(g::EGraph, r::SymbolicRule, id::Int64, matches::MatchesBuf)
+function search_rule!(g::EGraph, r::SymbolicRule, id::Int64, 
+    matches::MatchesBuf, mlock::ReentrantLock)
     for sub in ematch(g, r.left, id)
-        push!(matches, (r, r.right, sub, id))
+        lock(mlock) do
+            push!(matches, (r, r.right, sub, id))
+        end
     end
 end
 
-function search_rule!(g::EGraph, r::DynamicRule, id::Int64, matches::MatchesBuf)
+function search_rule!(g::EGraph, r::DynamicRule, id::Int64, 
+    matches::MatchesBuf, mlock::ReentrantLock)
     for sub in ematch(g, r.left, id)
-        push!(matches, (r, nothing, sub, id))
+        lock(mlock) do
+            push!(matches, (r, nothing, sub, id))
+        end
     end
 end
 
-function search_rule!(g::EGraph, r::BidirRule, id::Int64, matches::MatchesBuf)
+function search_rule!(g::EGraph, r::BidirRule, id::Int64, 
+    matches::MatchesBuf, mlock::ReentrantLock)
     for sub in ematch(g, r.left, id)
-        push!(matches, (r, r.right, sub, id))
+        lock(mlock) do
+            push!(matches, (r, r.right, sub, id))
+        end
     end
     for sub in ematch(g, r.right, id)
-        push!(matches, (r, r.left, sub, id))
+        lock(mlock) do
+            push!(matches, (r, r.left, sub, id))
+        end
     end
 end
 
@@ -51,7 +81,8 @@ function Base.show(io::IO, s::Sub)
     print(io, "]")
 end
 
-function search_rule!(g::EGraph, r::MultiPatRewriteRule, id::Int64, matches::MatchesBuf)
+function search_rule!(g::EGraph, r::MultiPatRewriteRule,
+    id::Int64, matches::MatchesBuf, mlock::ReentrantLock)
     buf = ematch(g, r.left, id)
     if isempty(buf)
         return 
@@ -75,15 +106,34 @@ function search_rule!(g::EGraph, r::MultiPatRewriteRule, id::Int64, matches::Mat
     end
     for sub in buf
[…]
