+++ paper/paper.bib
@@ -1,6 +1,6 @@
 @misc{rgf,
   author = {{Rackauckas}, C and {Foster}, C},
-  title = {RuntimeGeneratedFunctions.jl: Functions generated at runtime without world-age issues},
+  title = {{RuntimeGeneratedFunctions.jl}: Functions generated at runtime without world-age issues},
   year = {2021},
   publisher = {GitHub},
   journal = {GitHub repository},
@@ -9,7 +9,7 @@
 
 @misc{mlstyle,
   author = {{Zhao}, T},
-  title = {MLStyle.jl: Fast, consistent and extensible functional programming infrastructures},
+  title = {{MLStyle.jl}: Fast, consistent and extensible functional programming infrastructures},
   year = {2021},
   publisher = {GitHub},
   journal = {GitHub repository},
@@ -19,7 +19,7 @@
 
 @misc{metatheory,
   author = {{Cheli}, Alessandro},
-  title = {Metatheory.jl},
+  title = {{Metatheory.jl}},
   year = {2021},
   publisher = {GitHub},
   journal = {GitHub repository},
@@ -28,7 +28,7 @@
 
 @misc{matchcore,
   author = {{Zhao}, T and {Carlsson}, K},
-  title = {MatchCore.jl: A minimal implementation of optimized pattern matching.},
+  title = {{MatchCore.jl}: A minimal implementation of optimized pattern matching.},
   year = {2020},
   publisher = {GitHub},
   journal = {GitHub repository},
@@ -38,7 +38,7 @@
 
 @misc{datastruct,
   author = {{JuliaCollections}},
-  title = {DataStructures.jl: Implementation for a variety of data structures.},
+  title = {{DataStructures.jl}: Implementation for a variety of data structures.},
   year = {2021},
   publisher = {GitHub},
   journal = {GitHub repository},
@@ -91,7 +91,7 @@ eprint = {
 
 @article{belyakova2020world,
 author = {Belyakova, Julia and Chung, Benjamin and Gelinas, Jack and Nash, Jameson and Tate, Ross and Vitek, Jan},
-title = {World Age in Julia: Optimizing Method Dispatch in the Presence of Eval},
+title = {World Age in {Julia}: Optimizing Method Dispatch in the Presence of Eval},
 year = {2020},
 issue_date = {November 2020},
 publisher = {Association for Computing Machinery},
@@ -110,7 +110,7 @@ keywords = {compilation, dynamic languages, eval, method dispatch}
 
 @article{zappa2018julia,
 author = {Zappa Nardelli, Francesco and Belyakova, Julia and Pelenitsyn, Artem and Chung, Benjamin and Bezanson, Jeff and Vitek, Jan},
-title = {Julia Subtyping: A Rational Reconstruction},
+title = {{Julia} Subtyping: A Rational Reconstruction},
 year = {2018},
 issue_date = {November 2018},
 publisher = {Association for Computing Machinery},
