+++ paper/paper.bib
@@ -196,7 +196,7 @@ numpages = {9}
 }
 
 @article{philzuck1,
-  title={E-graphs in Julia {Part I}},
+  title={E-graphs in {Julia} {Part I}},
   author={Zucker, Philip},
   year={2020},
   url = {https://www.philipzucker.com/egraph-1/}
