### 2021-03-31 9cd40bd
finalizing custom types interface
