### 2021-03-31 2e39041
finalizing custom types interface
